### 2025-10-07 #9163 @editorialbot
[PRE REVIEW]: PSD: Parallel finite element Solver for continuum Dynamics
**Submitting author:** @mohd-afeef-badri (<a href="http://orcid.org/0000-0002-4646-8309">Mohd Afeef Badri</a>)
**Repository:** https://github.com/mohd-afeef-badri/psd
**Branch with paper.md** (empty if default branch):
**Version:** v2.6
