### 2025-10-07 #9163 @danielskatz
👋 @mohd-afeef-badri - Thanks for your submission.

However, it's far too long. JOSS suggests papers be 250-1000 words, and yours is over 2000. Please remove some of the text, perhaps by pointing to sections of the repo/docs from the paper. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. In addition, the command `@editorialbot check repository` will check a bunch of things, including the length of the paper. An editorialbot command needs to be the first entry in a new comment.

In addition, you could work on the invalid DOI that editorialbot notes. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.
