+++ joss/paper.bib
@@ -162,7 +162,7 @@ year = {2011}
   publisher={Blackwell Publishing Ltd Oxford, UK}
 }
 @inproceedings{cui2013physics,
-  doi = {110.1145/2503210.2503300 },
+  doi = {10.1145/2503210.2503300 },
   title={Physics-based seismic hazard analysis on petascale heterogeneous supercomputers},
   author={Cui, Yifeng and Poyraz, Efecan and Olsen, Kim B and Zhou, Jun and Withers, Kyle and Callaghan, Scott and Larkin, Jeff and Guest, C and Choi, D and Chourasia, Amit and others},
   booktitle={Proceedings of the international conference on high performance computing, networking, storage and analysis},
