### 2025-10-14 84643b8
short paper
