+++ joss/paper.md
@@ -30,39 +30,37 @@ bibliography: paper.bib
 
 # Summary
 
-`PSD` (Parallel finite element Solver for continuum Dynamics) is an open-source finite element solver designed for high-performance computing simulations in continuum dynamics with a special focus on earthquake mechanics and structural dynamics. `PSD` addresses the computational challenges of large-scale seismic simulations by providing an integrated platform for analyzing complex dynamics problems in both two and three dimensions.
+`PSD` (Parallel finite element Solver for continuum Dynamics) is an open-source finite element method (FEM) solver designed for high-performance computing simulations in continuum dynamics with a special focus on earthquake mechanics and structural dynamics. It enables integrated fault-to-site seismic simulations by combining advanced material modeling, scalable parallelism, and purpose-built meshing-partitioning tools.
 
-Built upon `FreeFEM`  [@MR3043640]  for finite element discretization and `PETSc` [@balay2019petsc] for scalable linear system solving, `PSD` integrates sophisticated material modeling through its dedicated  `Mfront` [@helfer2015],[@helfer2020] interface.   This architecture enables the handling of complex material behaviors thanks to non-linear constitutive laws essential for realistic mechanical phenomena modeling.   `PSD` includes a purpose-built `MPI I/O`-based mesher-partitioner, `top-ii-vol` [@badri2024top], tailored for large-scale earthquake simulations. Further, hybrid phase-field fracture mechanics [@ambati2015review] is implemented, which enables detailed analysis of crack initiation and propagation in materials. These types of simulations are treated at the other end of the complete simulation chain from the earthquake source to the structure assessment.
+Built upon `FreeFEM`  [@MR3043640]  for FEM discretization and `PETSc` [@balay2019petsc] for scalable linear system solving, `PSD` integrates sophisticated material modeling through its dedicated  `Mfront` [@helfer2015],[@helfer2020] interface, supporting non-linear constitutive laws for realistic simulations.   features a custom `MPI I/O`-based mesher-partitioner, `top-ii-vol` [@badri2024top], designed for earthquake scenarios and implements hybrid phase-field fracture mechanics [@ambati2015review] for crack analysis. This spans the full simulation chain from earthquake source to structural assessment.
 
-A distinguishing feature of `PSD` is its ability to handle large-scale earthquake simulations ($\mathcal{O}(10^9)$), particularly in fault-to-site analysis scenarios. This capability, combined with the solver's parallel architecture and advanced material modeling, makes it particularly suitable for comprehensive seismic risk assessment studies. The solver exhibits excellent scalability on tens of thousands of `MPI` processes, enabling simulations with billions of unknowns on large scale supercomputers. This allows it to accurately capture the multiscale features of seismic wave propagation and consequences—from fault to local site response.
+A key feature of `PSD` is its proven ability to handle fault-to-site earthquake simulations with billions of degrees of freedom, exhibiting excellent scalability on tens of thousands of MPI processes. This makes it a powerful tool for comprehensive seismic risk assessment.
 
 # Statement of Need
 
[…]
