### 2025-10-14 5c2fcec
short paper
