+++ joss/paper.md
@@ -38,11 +38,11 @@ A key feature of `PSD` is its proven ability to handle fault-to-site earthquake 
 
 # Statement of Need
 
-Seismic risk assessment require computational tools capable of simulating wave propagation across multiple spatial scales, from fault (kilometers away) to local site response (meters), while ensuring sufficient accuracy for earthquake engineering analyses. Existing commercial solutions often lack the computational scalability required for integrated regional to local scale earthquake simulations, while open-source alternatives typically address only specific aspects of the seismic simulation workflow.
+Seismic risk assessments require computational tools capable of simulating wave propagation across multiple spatial scales, from fault (kilometers away) to local site response (meters), while ensuring sufficient accuracy for earthquake engineering analyses. Existing commercial solutions often lack the computational scalability required for integrated regional earthquake simulations, while open-source alternatives typically address only specific aspects of the seismic simulation workflow.
 
-Current computational challenges in earthquake simulation include: (1) the need for billions of degrees-of-freedom to capture realistic fault-to-site scenarios [@hori2018application], [@cui2013physics] (2) integration of complex non-linear material behaviors and damage assessment for solids and structures, and (3) efficient mesh generation and partitioning for irregular geological domains often constructed directly from digital elevation models. Tools like `OpenSees`[@mckenna2011opensees] excel in local site response analysis (soils and structures) based on FEM and `SPECFEM3D` [@Peter_Forward_and_adjoint_2011] or `SEM3D` [@touhami2022sem3d] address seismic wave propagation using the spectral element method. However, it is of interest  to have an open-source integrated platform addressing the complete fault-to-site simulation workflow with HPC scalability.
+Current computational challenges in earthquake simulation include (1) the need for billions of degrees-of-freedom to capture realistic fault-to-site scenarios [@hori2018application], [@cui2013physics] (2) integration of complex non-linear material behaviors and damage assessment for solids and structures, and (3) efficient mesh generation and partitioning for irregular geological domains often constructed directly from digital elevation models. Tools like `OpenSees`[@mckenna2011opensees] excel in local site response analysis (soils and structures) based on FEM and `SPECFEM3D` [@Peter_Forward_and_adjoint_2011] or `SEM3D` [@touhami2022sem3d] address seismic wave propagation using the spectral element method. However, it is of interest  to have an open-source integrated platform addressing the complete fault-to-site simulation workflow with HPC scalability.
 
-`PSD` tries to fill this gap by providing a unified computational framework that combines earthquake source modeling, wave propagation simulation, and structural mechanics assessment  within a single scalable FEM solver.  `PDS`'s integration of advanced meshing-partitioning capabilities (`top-ii-vol`), sophisticated material modeling (`Mfront` interface), and fracture mechanics positions it uniquely for comprehensive seismic risk assessment needs requiring both regional-scale wave propagation and local site response including structural analysis.
[…]
