### 2025-10-16 339132f
shorten paper
