+++ joss/paper.bib
@@ -25,18 +25,6 @@
   year={2019},
   publisher={Argonne National Laboratory}
 }
-@article{helfer2020,
-  doi = {10.21105/joss.02003},
-  url = {https://doi.org/10.21105/joss.02003},
-  year = {2020},
-  publisher = {The Open Journal},
-  volume = {5},
-  number = {48},
-  pages = {2003},
-  author = {Thomas Helfer and Jeremy Bleyer and Tero Frondelius and Ivan Yashchuk and Thomas Nagel and Dmitri Naumov},
-  title = {The `MFrontGenericInterfaceSupport` project},
-  journal = {Journal of Open Source Software}
-}
 @article{helfer2015,
 	title = {Introducing the open-source mfront code generator: Application to mechanical behaviours and material knowledge management within the {PLEIADES} fuel element modelling platform},
 	volume = {70},
