+++ joss/paper.md
@@ -30,38 +30,35 @@ bibliography: paper.bib
 
 # Summary
 
-`PSD` (Parallel finite element Solver for continuum Dynamics) is an open-source finite element method (FEM) solver designed for high-performance computing simulations in continuum dynamics with a special focus on earthquake mechanics and structural dynamics. It enables integrated fault-to-site seismic simulations by combining advanced material modeling, scalable parallelism, and purpose-built meshing-partitioning tools.
+`PSD` (Parallel finite element Solver for continuum Dynamics) is an open-source finite element method (FEM) solver designed for high-performance computing simulations in continuum dynamics with a special focus on earthquake mechanics. It enables integrated fault-to-site seismic simulations by combining advanced material modeling, scalable parallelism, and purpose-built meshing-partitioning tools.
 
-Built upon `FreeFEM`  [@MR3043640]  for FEM discretization and `PETSc` [@balay2019petsc] for scalable linear system solving, `PSD` integrates sophisticated material modeling through its dedicated  `Mfront` [@helfer2015],[@helfer2020] interface, supporting non-linear constitutive laws for realistic simulations.   features a custom `MPI I/O`-based mesher-partitioner, `top-ii-vol` [@badri2024top], designed for earthquake scenarios and implements hybrid phase-field fracture mechanics [@ambati2015review] for crack analysis. This spans the full simulation chain from earthquake source to structural assessment.
+Built upon `FreeFEM`  [@MR3043640]  for FEM discretization and `PETSc` [@balay2019petsc] for scalable linear system solving, `PSD` integrates sophisticated non-linear material modeling through its  `Mfront` [@helfer2015] interface for realistic simulations.   It features a custom `MPI I/O`-based mesher-partitioner, `top-ii-vol` [@badri2024top] which provides efficient on-the-fly mesh generation and partitioning specifically designed for earthquake simulation geometries, eliminating traditional bottlenecks associated with sequential meshing approaches.  On the structural side, hybrid phase-field fracture mechanics [@ambati2015review] for crack analysis is implemented. This spans the full simulation chain from earthquake source to structural assessment.
 
-A key feature of `PSD` is its proven ability to handle fault-to-site earthquake simulations with billions of degrees of freedom, exhibiting excellent scalability on tens of thousands of MPI processes. This makes it a powerful tool for comprehensive seismic risk assessment.
+A key feature of `PSD` is its proven ability to handle fault-to-site earthquake simulations with billions of degrees of freedom, exhibiting scalability on tens of thousands of MPI processes, making it a tool for comprehensive seismic risk assessment.
 
 # Statement of Need
 
-Seismic risk assessments require computational tools capable of simulating wave propagation across multiple spatial scales, from fault (kilometers away) to local site response (meters), while ensuring sufficient accuracy for earthquake engineering analyses. Existing commercial solutions often lack the computational scalability required for integrated regional earthquake simulations, while open-source alternatives typically address only specific aspects of the seismic simulation workflow.
[…]
