### 2025-10-17 13a181c
reduce word count
