+++ joss/paper.md
@@ -32,17 +32,17 @@ bibliography: paper.bib
 
 `PSD` (Parallel finite element Solver for continuum Dynamics) is an open-source finite element method (FEM) solver designed for high-performance computing simulations in continuum dynamics with a special focus on earthquake mechanics. It enables integrated fault-to-site seismic simulations by combining advanced material modeling, scalable parallelism, and purpose-built meshing-partitioning tools.
 
-Built upon `FreeFEM`  [@MR3043640]  for FEM discretization and `PETSc` [@balay2019petsc] for scalable linear system solving, `PSD` integrates sophisticated non-linear material modeling through its  `Mfront` [@helfer2015] interface for realistic simulations.   It features a custom `MPI I/O`-based mesher-partitioner, `top-ii-vol` [@badri2024top] which provides efficient on-the-fly mesh generation and partitioning specifically designed for earthquake simulation geometries, eliminating traditional bottlenecks associated with sequential meshing approaches.  On the structural side, hybrid phase-field fracture mechanics [@ambati2015review] for crack analysis is implemented. This spans the full simulation chain from earthquake source to structural assessment.
+Built on `FreeFEM` [@MR3043640] for FEM discretization and `PETSc` [@balay2019petsc] for scalable linear solvers, `PSD` integrates non-linear material modeling through its `MFront` [@helfer2015] interface for realistic simulations. Its custom `MPI I/O`-based mesher–partitioner, `top-ii-vol` [@badri2024top], enables efficient on-the-fly mesh generation and partitioning for earthquake geometries, removing sequential meshing bottlenecks. On the structural side, hybrid phase-field fracture mechanics [@ambati2015review] for crack analysis is implemented. This spans the full simulation chain from earthquake source to structural assessment.
 
-A key feature of `PSD` is its proven ability to handle fault-to-site earthquake simulations with billions of degrees of freedom, exhibiting scalability on tens of thousands of MPI processes, making it a tool for comprehensive seismic risk assessment.
+A key feature of `PSD` is its ability to perform fault-to-site earthquake simulations with billions of degrees of freedom, scaling efficiently on tens of thousands of `MPI`-processes, making it suitable for comprehensive seismic risk assessment.
 
 # Statement of Need
 
-Seismic risk assessments require computational tools capable of simulating wave propagation across multiple spatial scales, from fault (kilometers away) to local site response (meters), while ensuring sufficient accuracy for earthquake analyses. Existing commercial solutions often lack the computational scalability required for integrated regional earthquake simulations, while open-source alternatives typically address specific aspects of the seismic simulation workflow.
+Seismic risk assessment requires tools that can simulate wave propagation across multiple scales, from faults (kilometers away) to local site response (meters), with sufficient accuracy. Commercial software often lacks the scalability needed for regional simulations, while open-source alternatives typically focus on specific parts of the seismic workflow.
 
[…]
