### 2025-10-17 aece710
refer to maths in doc
