+++ joss/paper.md
@@ -40,19 +40,19 @@ A key feature of `PSD` is its ability to perform fault-to-site earthquake simula
 
 Seismic risk assessment requires tools that can simulate wave propagation across multiple scales, from faults (kilometers away) to local site response (meters), with sufficient accuracy. Commercial software often lacks the scalability needed for regional simulations, while open-source alternatives typically focus on specific parts of the seismic workflow.
 
-Current computational challenges in earthquake simulation include: (1) handling billions of degrees-of-freedom to capture realistic fault-to-site scenarios [@hori2018application], [@cui2013physics], (2) integrating complex non-linear material behaviors and damage assessment for solids and structures, and (3) efficiently generating and partitioning large meshes for irregular geological domains derived from digital elevation models. FEM tools like `OpenSees`[@mckenna2011opensees] excel in local-site response, while `SPECFEM3D` [@Peter_Forward_and_adjoint_2011] and `SEM3D` [@touhami2022sem3d] address wave propagation using spectral elements. An open-source platform covering the full fault-to-site workflow with HPC scalability remains highly desirable.
+Current computational challenges in earthquake simulation include: (1) handling billions of degrees-of-freedom to capture realistic fault-to-site scenarios [@hori2018application],[@cui2013physics], (2) integrating complex non-linear material behaviors and damage assessment for solids and structures, and (3) efficiently generating and partitioning large meshes for irregular geological domains derived from digital elevation models. FEM tools like `OpenSees`[@mckenna2011opensees] excel in local-site response, while `SPECFEM3D` [@Peter_Forward_and_adjoint_2011] and `SEM3D` [@touhami2022sem3d] address wave propagation using spectral elements. An open-source platform covering the full fault-to-site workflow with HPC scalability remains highly desirable.
 
 `PSD` tries to fill this gap by providing a unified framework that combines earthquake wave propagation simulation and structural mechanics assessment  within a single scalable FEM solver.  `PDS`'s integration of advanced meshing-partitioning capabilities, sophisticated material modeling, and fracture mechanics positions it uniquely for comprehensive seismic risk assessment needs requiring both regional-scale wave propagation and local site response, including structural analysis.
 
 # Features and Architecture
 
-`PSD` offers a range of physics modules designed for earthquake simulations, including linear elasticity, elastodynamics, fracture mechanics, soildynamics [^soildynamics], and elasto-plasticity. Its versatility is further enhanced by the `Mfront` interface, which allows users to implement custom non-linear material models or choose from hundreds of models, that can be seamlessly integrated into any of these modules expanding their capabilities beyond its built-in constitutive laws. Additionally, comprehensive verification and validation campaigns for all `PSD` modules, with results cross-compared against reference codes, experimental data, and analytical benchmarks, ensuring transparency and reproducibility (see [validation page](https://mohd-afeef-badri.github.io/psd/validation)).
[…]
