### 2025-10-17 #9163 @mohd-afeef-badri
Dear @danielskatz,

Thank you for your helpful feedback.

I have carefully addressed the issues you noted. The incorrect DOI has been fixed and verified, and the manuscript has been substantially revised to meet the JOSS length requirements, now totaling approximately 1000 words (excluding metadata). I have also reviewed the title as you suggested and overall formatting for clarity.

I hope the revised submission satisfies the criteria for proceeding to the next stage of the review process.
