### 2025-11-24 mohd-afeef-badri/psd#56 @gtheler
JOSS review
Hi team. I will be reviewing your JOSS paper so I opened this isse to start the discuss a few things.

First, +1 for using autotools instead of cmake.
I get a warning about a deprecated macro:

```
$ autoreconf -i
configure.ac:294: warning: The macro 'AC_TRY_COMPILE' is obsolete.
configure.ac:294: You should run autoupdate.
./lib/autoconf/general.m4:2845: AC_TRY_COMPILE is expanded from...
configure.ac:181: AX_MPI is expanded from...
configure.ac:294: the top level
configure.ac:299: warning: The macro 'AC_TRY_COMPILE' is obsolete.
configure.ac:299: You should run autoupdate.
./lib/autoconf/general.m4:2845: AC_TRY_COMPILE is expanded from...
configure.ac:181: AX_MPI is expanded from...
configure.ac:299: the top level
configure.ac:297: installing './compile'
configure.ac:2: installing './install-sh'
configure.ac:2: installing './missing'
```

Then, when configuring using `--with-dependencies=yes` I get a weird error when compiling metis inside PETSc:

```
cd /home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/petsc-build/libmetis && /usr/bin/gcc -Dmetis_EXPORTS -I/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/GKlib -I/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/include -I/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/libmetis/. -I/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/petsc-build/include -fPIC -Wno-lto-type-mismatch -Wno-stringop-overflow -Wno-lto-type-mismatch -Wno-stringop-overflow -g -O2 -DLINUX -D_FILE_OFFSET_BITS=64 -std=c99 -fno-strict-aliasing -fPIC -Wall -pedantic -Wno-unused-variable -Wno-unknown-pragmas -DNDEBUG -DNDEBUG2 -DUSE_GKRAND -DHAVE_EXECINFO_H -DHAVE_GETLINE  -fPIC -Wno-lto-type-mismatch -Wno-stringop-overflow -Wno-lto-type-mismatch -Wno-stringop-overflow -g -O2 -fPIC -MD -MT libmetis/CMakeFiles/metis.dir/__/GKlib/fs.c.o -MF CMakeFiles/metis.dir/__/GKlib/fs.c.o.d -o CMakeFiles/metis.dir/__/GKlib/fs.c.o -c /home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/GKlib/fs.c
gmake[5]: Leaving directory '/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/petsc-build'
gmake[4]: Leaving directory '/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/petsc-build'
gmake[3]: Leaving directory '/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/petsc-build'
stdout:
/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/GKlib/error.c: In function ‘gk_strerror’:
/home/gtheler/codigos/psd/ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/GKlib/error.c:183:3: error: implicit declaration of function ‘strerror_r’; did you mean ‘strerror’? [-Wimplicit-function-declaration]
  183 |   strerror_r(errnum, buf, 1024);
      |   ^~~~~~~~~~
      |   strerror
gmake[5]: *** [libmetis/CMakeFiles/metis.dir/build.make:124: libmetis/CMakeFiles/metis.dir/__/GKlib/error.c.o] Error 1
gmake[5]: *** Waiting for unfinished jobs....
```

I am not sure why it does not have `strerror_r` since it is pretty standard.
Have you seen this?

I already have PETSc in my own `PETSC_DIR` with `PETSC_ARCH`. Can use it?
I do not have the rest of the dependencis. Is there a way to have configure install all the other dependencies but use my own PETSc (or even petsc from apt?)
