### 2025-11-25 9b6230b
update the minimal version should resolv the warning about a deprecated macro issue #56
