### 2025-11-25 mohd-afeef-badri/psd#56 @mohd-afeef-badri
@gtheler hello,

Thanks a lot for this first feedback and for starting the review process! Sorry that the installation did not go through as expected.

**Regarding Autotools warning**

I have now updated the Autoconf version ([9b6230b](https://github.com/mohd-afeef-badri/psd/commit/9b6230bd65e4a13e86278bf86365e541a1af3dad)) used to generate the build system, and this removes the warnings related to the deprecated `AC_TRY_COMPILE` macro. The next push  will include regenerated `configure` files so those warnings should disappear on your side as well.

**Regarding METIS / `strerror_r` compilation error**

Sadly, we have not seen this particular `strerror_r` implicit-declaration error before. In our CI (https://github.com/mohd-afeef-badri/psd/tree/main/.github/workflows) the full dependency chain (MPICH, OpenMPI, PETSc, FreeFEM, METIS/ParMETIS, etc.) builds cleanly. You can check the exact workflow here, which shows the full pipeline and tested environment:

To better understand the issue on our side :
**Which operating system and version are you building on?**

This will maybe  help us reproduce the missing `strerror_r` problem.

**Using your own alternative installation route**

There is an alternative way to install PSD this is refered to as **Procedure 2 in the documentation** (https://mohd-afeef-badri.github.io/psd/#/install?id=installation-procedure-2). In your case we recommend this approach, because it avoids rebuilding the full PETSc/METIS/FreeFEM toolchain.

The most reliable setup is:

1.Install Gmsh via the official tarball, and
2. Use FreeFEM binaries, which already come with a consistent PETSc, METIS, ParMETIS, MUMPS, SCOTCH, etc. pre-packaged:
   [https://github.com/FreeFem/FreeFem-sources/releases/tag/v4.15](https://github.com/FreeFem/FreeFem-sources/releases/tag/v4.15)

If `FreeFem` is in the `PATH`, PSD will detect it automatically.
Alternatively, you can specify paths explicitly:

```
./configure --with-FreeFEM=/path/to/FreeFEM \
            --with-Gmsh=/path/to/gmsh
```

(`./configure --help` lists more options.)

Hope this answer helps to resolve the issue.

In any case, we are looking forward to working with you to resolve all of this, and thank you again for the careful review!
