### 2025-12-11 mohd-afeef-badri/psd#56 @gtheler
I tweaked some things and made it compile. Now `make check` complains but I don't see `test-suite.log`:

```
gtheler@tom:~/codigos/psd$ make check
Making check in ext
make[1]: Entering directory '/home/gtheler/codigos/psd/ext'

make  check-local
make[2]: Entering directory '/home/gtheler/codigos/psd/ext'

*============================================================*
 Nothing to check here:
*============================================================*

make[2]: Leaving directory '/home/gtheler/codigos/psd/ext'
make[1]: Leaving directory '/home/gtheler/codigos/psd/ext'
Making check in src
make[1]: Entering directory '/home/gtheler/codigos/psd/src'
Making check in plugins
make[2]: Entering directory '/home/gtheler/codigos/psd/src/plugins'
Making check in top-ii-vol
make[3]: Entering directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol'
Making check in cpp
make[4]: Entering directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol/cpp'

*============================================================*
 Compilation done in :
/home/gtheler/codigos/psd/src/plugins/top-ii-vol/cpp
*============================================================*

 Compilation list top-ii-vol.cpp
make[4]: Leaving directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol/cpp'
Making check in test
make[4]: Entering directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol/test'

make  check-local
make[5]: Entering directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol/test'

*============================================================*
 Checking top-ii-vol plugin : RUNNING
*============================================================*

Test_1 : top-ii-vol load test
  PASSED
Test_2 :  top-ii-vol testing point-cloud  partitioning
  PASSED
Test_3 :  top-ii-vol testing point-cloud meshing and partitioning
  PASSED
Test_4 :  top-ii-vol testing point-cloud meshing and X-partitioning
--------------------------------------------------------------------------
prterun detected that one or more processes exited with non-zero status,
thus causing the job to be terminated. The first process to do so was:

   Process name: [prterun-tom-151424@1,0]
   Exit code:    2
--------------------------------------------------------------------------
make[5]: *** [Makefile:470: Test_4] Error 2
make[5]: Leaving directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol/test'
make[4]: *** [Makefile:317: check-am] Error 2
make[4]: Leaving directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol/test'
make[3]: *** [Makefile:344: check-recursive] Error 1
make[3]: Leaving directory '/home/gtheler/codigos/psd/src/plugins/top-ii-vol'
make[2]: *** [Makefile:344: check-recursive] Error 1
make[2]: Leaving directory '/home/gtheler/codigos/psd/src/plugins'
make[1]: *** [Makefile:354: check-recursive] Error 1
make[1]: Leaving directory '/home/gtheler/codigos/psd/src'
make: *** [Makefile:410: check-recursive] Error 1
```
