### 2025-12-11 mohd-afeef-badri/psd#56 @mohd-afeef-badri
@gtheler thank you for looking into the problem, I would definitely be interested in the tweak you made for the installation, perhaps this will help to improve PSD so that such installation issues are not faced.

Concerning the failing test you will find the log file `Test_4.log` in the folder where the test failed. In your case it will be `/home/gtheler/codigos/psd/src/plugins/top-ii-vol/test/Test_4.log`. Can you please share this file here so that I can have a look why it fails.

Also, as the location of log file is not intuitive, I will try to add a remark on log file locations in the documentation.
