### 2025-12-17 bb0fa4b
better installation support document
