+++ docs/install.md
@@ -1,290 +1,181 @@
-**PSD Platform Support and Installation Guide**
+PSD is a cross‑platform, MPI‑enabled high‑performance finite element solver (FEM) for Unix/Linux systems. This installation guide describes required dependencies, step‑by‑step instructions to build from source (Ubuntu/macOS examples), how to use the official Docker image, and the available `./configure` and `make` options.
 
-**PSD** is a cross-platform finite element method (FEM) solver primarily designed for Unix/Linux environments. It has been successfully deployed on the following systems:
+  ---
 
-* **CentOS**: 7 / 8
-* **Ubuntu**: 16.04 / 18.04 / 20.04 / 22.04 / 24.04
-* **Debian**: 11
-* **Raspberry Pi OS**
-* **Fedora**: 30 / 32 / 34 / 36 / 38
-* **macOS**:
-  * Intel: Puma 10.1.5
-  * ARM (M2): Ventura 13
+  **Prerequisites**
 
----
+  Before installing PSD, ensure the following dependencies are installed and compatible on your system.
 
-**Prerequisites**
+  | Package                                                      | Required Version                                             | Essential | Install Assist |
+  | ------------------------------------------------------------ | ------------------------------------------------------------ | --------- | -------------- |
+  | [automake](https://www.gnu.org/software/automake/)           | 2.8 or higher                                                | ✅         | ❌              |
+  | [C/C++ compiler](http://www.cplusplus.com/)                  | GCC 7 or higher                                              | ✅         | ❌              |
+  | [MPI](https://www.mpich.org/)                                | 2.0 or higher ([MPICH](https://www.mpich.org/) or [Open MPI](https://www.open-mpi.org/)) | ✅         | ❌              |
+  | [git](https://git-scm.com/)                                  | –                                                            | ✅         | ❌              |
+  | [FreeFEM](https://freefem.org/)                              | 4.15                                                         | ✅         | ✅              |
+  | [PETSc](https://www.mcs.anl.gov/petsc/)                      | 3.22.2                                                       | ✅         | ✅              |
+  | [Gmsh](http://gmsh.info/)                                    | 4.11.1                                                       | ✅         | ✅              |
+  | [SALOME](https://www.salome-platform.org/)                   | 9.0 or higher                                                | ❌         | ❌              |
+  | [gnuplot](http://www.gnuplot.info/)                          | 4.0 or higher                                                | ❌         | ❌              |
+  | [MFront](http://tfel.sourceforge.net/)                       | 5.0.0                                                        | ❌         | ✅              |
+  | [MGIS](https://thelfer.github.io/mgis/web/bindings-cxx.html) | 3.0                                                          | ❌         | ✅              |
 
-Before installing PSD, ensure the following dependencies are installed and compatible on your system.
+  > ⚠️ **Note:**
+  > Some dependencies can be automatically installed with PSD. These are marked in the “Install Assist” column above. Others must be installed manually before proceeding.
 
-| Package                                                   | Required Version                                                                   | Essential | Install Assist |
-| --------------------------------------------------------- | ---------------------------------------------------------------------------------- | --------- | -------------- |
-| [automake](https://www.gnu.org/software/automake/)           | 2.8 or higher                                                                      | ✅        | ❌             |
[…]
