### 2025-12-17 d812211
add Automakes standard test logic
