### 2025-12-17 0c3408b
mention Windows usage
