### 2025-12-17 mohd-afeef-badri/psd#56 @mohd-afeef-badri
@gtheler Thank you very much for taking the time to investigate the installation issues so thoroughly, and for sharing your experience and suggestions, I really appreciate the detailed feedback. Really sorry that the installation did not go through so smoothly.

You are absolutely right that, in a classic Unix setting, relying on system-provided dependencies (apt/yum..) greatly reduces surprises. That said, the installation strategy used by PSD is intentionally different, and this is largely driven by its primary usage context. PSD is mostly deployed on HPC clusters, where internet access is often restricted or unavailable on compute nodes, and users do not have root privileges.

For this reason, PSD follows a self-contained download-and-build principle, where external dependencies (e.g. PETSc, FreeFEM plugins) are built in a controlled environment using the same MPI toolchain as the target platform. While this can be less aligned with the "least surprise" rule on a developer workstation, it has proven to be more robust and reproducible on HPC systems which we have tested.

In all cases, thanks to your remarks to mitigate the complexity you encountered and to facilitate both user onboarding and the JOSS review process, I have provided a ready-to-use Docker image that contains a fully functional PSD installation, including FreeFEM, PETSc, MPI, and all required plugins, you can find more info [here](https://mohd-afeef-badri.github.io/psd/#/install?id=docker-image-for-psd).

Finally,  your point about clearer diagnostics (e.g. via `test-suite.log`) is well taken, folllowing your advice I have now moved PSD's testing to more standard Atomake testing that produces the standard `test-suite.log` file upon failures (implemented via [commit](https://github.com/mohd-afeef-badri/psd/commit/d8122115b9311be86468bc6ea66120588fde0479)). Thank you again for your review.

I hope that the docker based procedure helps, additionally, I have also pushed the [Dockerfile](https://github.com/mohd-afeef-badri/psd/blob/main/Dockerfile) that helps build PSD from AtoZ on a bare Ubuntu 24.04 machine. And I have personally tested this on on Linux (Ubuntu 24.04/ Fedora 40), MacOS (Tahoe 26.2) and Windows 11 verified that it works.
