### 2025-12-23 mohd-afeef-badri/psd#56 @mohd-afeef-badri
Ahh this is better now I can see the error is due to `lmpi_cxx` not present or not being detected by FreeFEM.
can you please verify with
```
ldd /usr/local/bin/FreeFem++-mpi
```
that `libmpi_cxx` is found for your system ?
