### 2025-12-24 mohd-afeef-badri/psd#56 @gtheler
So we need a "new" FreeFEM but not too new OpenMPI. A litte bit contradicting, isn't it?

I think I said it before but I do not like Docker nor Ubuntu. I use Debian. I understand if you target Ubuntu, you are free to say "we only support Ubuntu." Not that I agree, but I understand it.
I also do not like OpenMPI (I prefer MPICH) but that's what Paraview and all those guys are linked against in Debian, so that's why OpenMPI 5 is there. Because I use the versions from `apt`.

I started all over from scratch.
I had to do this:

1. unset existing `PETSC_ARCH` and `PETSC_DIR` (can psd use an exsiting petsc installation?)
2. remove petsc cmake patch

```
diff --git a/ext/Makefile.am b/ext/Makefile.am
index e207356..f46e104 100644
--- a/ext/Makefile.am
+++ b/ext/Makefile.am
@@ -148,7 +148,7 @@ PTSCOTCH_DOWNLOAD = wget https://gitlab.inria.fr/scotch/scotch/-/archive/v$(PTSC
 MSHMET_DOWNLOAD = wget https://raw.githubusercontent.com/FreeFem/FreeFEM-3rdparties/refs/heads/main/mshmet.$(MSHMET_VERSION).tgz \
 			-O mshmet.$(MSHMET_VERSION).tgz
 # Download PETSc patch for cmake
-PETSC_PATCH_DOWNLOAD = wget https://raw.githubusercontent.com/FreeFem/FreeFem-sources/refs/tags/v$(FREEFEM_VERSION)/3rdparty/ff-petsc/petsc-cmake-seq.patch
+# PETSC_PATCH_DOWNLOAD = wget https://raw.githubusercontent.com/FreeFem/FreeFem-sources/refs/tags/v$(FREEFEM_VERSION)/3rdparty/ff-petsc/petsc-cmake-seq.patch
 # Download Ipopt source code
 IPOPT_DOWNLOAD = wget -c http://www.coin-or.org/download/source/Ipopt/Ipopt-$(IPOPT_VERSION).tgz \
 			-O Ipopt-$(IPOPT_VERSION).tgz
@@ -232,8 +232,8 @@ download-dependencies:
 	@if [ ! -f f2cblaslapack-$(F2CBLASLAPACK_VERSION).tar.gz ]; then \
 		$(F2CBLASLAPACK_DOWNLOAD); \
 	fi
-	@if [ ! -f petsc-cmake-seq.patch ]; then \
-		$(PETSC_PATCH_DOWNLOAD); \
+# 	@if [ ! -f petsc-cmake-seq.patch ]; then \
+# 		$(PETSC_PATCH_DOWNLOAD); \
 	fi
 	@if [ ! -f Ipopt-$(IPOPT_VERSION).tgz ]; then \
 		$(IPOPT_DOWNLOAD); \
@@ -453,7 +453,7 @@ endif

 petsc-build: petsc-$(PETSC_VERSION)/$(PETSC_ARCH_PSD)/lib/libpetsc.$(SO_SUFFIX)

-petsc-$(PETSC_VERSION)/$(PETSC_ARCH_PSD)/lib/libpetsc.$(SO_SUFFIX): petsc-cmake-seq.patch petsc-$(PETSC_VERSION).tar.gz
+petsc-$(PETSC_VERSION)/$(PETSC_ARCH_PSD)/lib/libpetsc.$(SO_SUFFIX): petsc-$(PETSC_VERSION).tar.gz
 	@echo ""
 	@echo "*============================================================*"
 	@echo " PSD is compiling PETSc, HPDDM, METIS, PARMETIS, SCALAPACK,   "
@@ -466,8 +466,7 @@ petsc-$(PETSC_VERSION)/$(PETSC_ARCH_PSD)/lib/libpetsc.$(SO_SUFFIX): petsc-cmake-
 	tar -xf petsc-$(PETSC_VERSION).tar.gz
 	[ -d petsc-v$(PETSC_VERSION) ] && mv petsc-v$(PETSC_VERSION) petsc-$(PETSC_VERSION)
 	cd petsc-$(PETSC_VERSION) && \
-	patch -p1 < ../petsc-cmake-seq.patch     && \
-	./configure \
+ 	./configure \
 	--with-scalar-type=real \
 	--prefix=${prefix}/ff-petsc/r \
 	--with-cudac=0 --with-ssl=0 --with-x=0 \
@@ -489,10 +488,10 @@ if !BUILDZIPPEDDEPENDENCIES
 	$(PETSC_DOWNLOAD)
 endif

-petsc-cmake-seq.patch:
-if !BUILDZIPPEDDEPENDENCIES
-	$(PETSC_PATCH_DOWNLOAD)
-endif
+# petsc-cmake-seq.patch:
+# if !BUILDZIPPEDDEPENDENCIES
+# 	$(PETSC_PATCH_DOWNLOAD)
+# endif

 #--------------------------------------------------------------------------------
@@ -662,8 +661,8 @@ maintainer-clean-local: clean-local
 	@echo " Cleaning automake generated files"
 	@echo "*============================================================*"
 	@echo ""
-	rm  -rf Makefile Makefile.in petsc-cmake-seq.patch
-	rm  -rf petsc-cmake-seq.patch
+	rm  -rf Makefile Makefile.in
+# 	rm  -rf petsc-cmake-seq.patch
 	rm  -rf petsc-$(PETSC_VERSION) freefem-$(FREEFEM_VERSION) mfront-$(MFRONT_VERSION)
 	rm  -rf mgis-$(MGIS_VERSION) gmsh-$(GMSH_VERSION) Ipopt-$(IPOPT_VERSION).tgz mshmet.$(MSHMET_VERSION).tgz
 	rm  -rf *.tar.gz
@@ -680,4 +679,4 @@ install-exec-local:
 	@echo "*============================================================*"
 	@echo " Nothing to install here:"
 	@echo "*============================================================*"
-	@echo ""
\ No newline at end of file
+	@echo ""
```

3. add

```
#define _GNU_SOURCE
#include <string.h>
```

to `ext/petsc-3.22.2/fr/externalpackages/petsc-pkg-metis-8b194fdf0966/GKlib/error.c`

I now get

```
make[5]: Entering directory '/home/gtheler/codigos/psd/src/plugins/medio/cpp'
/home/gtheler/PSDinstall/bin/ff-c++ -auto medio.cpp -I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/hdf5/include -I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/medfile/include -I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/INSTALL/MEDCOUPLING/include -g -O2  -L/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/hdf5/lib -L/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/medfile/lib -L/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/INSTALL/MEDCOUPLING/lib  -lhdf5 -lmedC -lmedloader -lmedcoupling -linterpkernel -lmedcouplingremapper   -DMEDCOUPLING
eval /home/gtheler/PSDinstall/bin/ff-c++ '-I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/hdf5/include' '-I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/medfile/include' '-I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/INSTALL/MEDCOUPLING/include' '-g' '-O2' '-L/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/hdf5/lib' '-L/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/medfile/lib' '-L/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/INSTALL/MEDCOUPLING/lib' '-lhdf5' '-lmedC' '-lmedloader' '-lmedcoupling' '-linterpkernel' '-lmedcouplingremapper' '-DMEDCOUPLING' medio.cpp
mpic++ -c -fPIC -g -O2 '-I/home/gtheler/PSDinstall/lib/ff++/4.15/include' '-I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/hdf5/include' '-I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/BINARIES-CO7/medfile/include' '-I/home/gtheler/PSDinstall/MEDCOUPLING-9.14.0/INSTALL/MEDCOUPLING/include' '-DMEDCOUPLING' -g -mmmx -mavx2 -mavx -msse4.2 -msse2 -msse -std=gnu++14 -DBAMG_LONG_LONG -DNCHECKPTR -fPIC -I/usr/include/suitesparse 'medio.cpp'
medio.cpp:27:10: fatal error: MEDLoader.hxx: No such file or directory
   27 | #include "MEDLoader.hxx"
      |          ^~~~~~~~~~~~~~~
compilation terminated.
make[5]: *** [Makefile:445: medio.so] Error 1
make[5]: Leaving directory '/home/gtheler/codigos/psd/src/plugins/medio/cpp'
make[4]: *** [Makefile:344: all-recursive] Error 1
make[4]: Leaving directory '/home/gtheler/codigos/psd/src/plugins/medio'
make[3]: *** [Makefile:344: all-recursive] Error 1
make[3]: Leaving directory '/home/gtheler/codigos/psd/src/plugins'
make[2]: *** [Makefile:354: all-recursive] Error 1
make[2]: Leaving directory '/home/gtheler/codigos/psd/src'
make[1]: *** [Makefile:410: all-recursive] Error 1
make[1]: Leaving directory '/home/gtheler/codigos/psd'
make: *** [Makefile:351: all] Error 2
```
