### 2026-01-06 #9411 @gtheler
> Hi [@gtheler](https://github.com/gtheler), [@chennachaos](https://github.com/chennachaos) how is your review going?

Still not able to compile: https://github.com/mohd-afeef-badri/psd/issues/56#issuecomment-3687654330
