### 2026-01-15 5626750
Update README with MacOs-26 CI badge
Added MacOs-26 badge to the README for CI status.
