### 2026-01-28 #9411 @gtheler
@diehlpk I could compile and run the tests but I cannot run my own cases to compare results with other existing software: https://github.com/mohd-afeef-badri/psd/issues/56#issuecomment-3770137725
