### 2026-02-02 mohd-afeef-badri/psd#56 @diehlpk
@mohd-afeef-badri just be aware the freefem++ spack package is broken and you would need to fix that too.
