### 2026-02-02 #9411 @diehlpk
@mohd-afeef-badri  it seems that the installation instructions are not very good. Maybe you could automatize the installation?
