### 2026-02-03 #9411 @mohd-afeef-badri
@diehlpk thanks for your message.

Just for context: I already provide [Docker images for PSD](https://mohd-afeef-badri.github.io/psd/#/install?id=docker-image-for-psd), which I can build successfully from a clean Ubuntu 24.04 or 22.04 environment. From my experience, the main difficulty is not the build itself, but rather the variability of users local installation environments. Moreover it is installing the dependencies correctly which is the issue.

In fact, the JOSS review process motivated me to create and document these Docker images and their recipes, which was not originally planned but has been a very useful outcome of the review.

As a possible next step toward a more automated and user-friendly installation, I am considering packaging PSD with Spack. My concern is mainly the additional time required, since I would first need to learn Spack and then prepare and maintain the PSD package.

From my own experience as a JOSS reviewer, Docker-based installation has been considered an acceptable alternative in similar cases [link](https://github.com/openjournals/joss-reviews/issues/7017) [link](https://github.com/openjournals/joss-reviews/issues/7017#issuecomment-2327505654) (where setup of dependencies was difficult). Do you think that the current Docker-based workflow would already be sufficient to address the installation concerns, or would you recommend adding a Spack package as a better (or complementary) solution?
