### 2026-02-06 #9411 @diehlpk
@mohd-afeef-badri I think it is fine to use a docker image for the reviewers. However, at least one should be able to compile your code.
