### 2026-02-07 mohd-afeef-badri/psd#56 @gtheler
I started over from scratch in a fresh laptop with Debian Trixie:

```
gtheler@chalmers:~/psd$ wget https://github.com/FreeFem/FreeFem-sources/releases/download/v4.15/FreeFEM-4.15-amd64-ubuntu24.04.deb
--2026-02-07 17:51:47--  https://github.com/FreeFem/FreeFem-sources/releases/download/v4.15/FreeFEM-4.15-amd64-ubuntu24.04.deb
Resolving github.com (github.com)... 4.228.31.150
Connecting to github.com (github.com)|4.228.31.150|:443... connected.
HTTP request sent, awaiting response... 302 Found
Location: https://release-assets.githubusercontent.com/github-production-release-asset/121615834/53c8c3ca-307a-4031-b51c-aa9c3264ce19?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-02-07T21%3A30%3A25Z&rscd=attachment%3B+filename%3DFreeFEM-4.15-amd64-ubuntu24.04.deb&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-02-07T20%3A30%3A25Z&ske=2026-02-07T21%3A30%3A25Z&sks=b&skv=2018-11-09&sig=fanS6zwOtPKZriHINCurXMm899fwo%2BPVOseVfM8PIRM%3D&jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmVsZWFzZS1hc3NldHMuZ2l0aHVidXNlcmNvbnRlbnQuY29tIiwia2V5Ijoia2V5MSIsImV4cCI6MTc3MDUwMTEwNywibmJmIjoxNzcwNDk3NTA3LCJwYXRoIjoicmVsZWFzZWFzc2V0cHJvZHVjdGlvbi5ibG9iLmNvcmUud2luZG93cy5uZXQifQ.J0uxjvn1lU0jCrtumqmlGoUJE2_TxJeC2RST4G-bHio&response-content-disposition=attachment%3B%20filename%3DFreeFEM-4.15-amd64-ubuntu24.04.deb&response-content-type=application%2Foctet-stream [following]
--2026-02-07 17:51:47--  https://release-assets.githubusercontent.com/github-production-release-asset/121615834/53c8c3ca-307a-4031-b51c-aa9c3264ce19?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-02-07T21%3A30%3A25Z&rscd=attachment%3B+filename%3DFreeFEM-4.15-amd64-ubuntu24.04.deb&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-02-07T20%3A30%3A25Z&ske=2026-02-07T21%3A30%3A25Z&sks=b&skv=2018-11-09&sig=fanS6zwOtPKZriHINCurXMm899fwo%2BPVOseVfM8PIRM%3D&jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmVsZWFzZS1hc3NldHMuZ2l0aHVidXNlcmNvbnRlbnQuY29tIiwia2V5Ijoia2V5MSIsImV4cCI6MTc3MDUwMTEwNywibmJmIjoxNzcwNDk3NTA3LCJwYXRoIjoicmVsZWFzZWFzc2V0cHJvZHVjdGlvbi5ibG9iLmNvcmUud2luZG93cy5uZXQifQ.J0uxjvn1lU0jCrtumqmlGoUJE2_TxJeC2RST4G-bHio&response-content-disposition=attachment%3B%20filename%3DFreeFEM-4.15-amd64-ubuntu24.04.deb&response-content-type=application%2Foctet-stream
Resolving release-assets.githubusercontent.com (release-assets.githubusercontent.com)... 185.199.111.133, 185.199.109.133, 185.199.108.133, ...
Connecting to release-assets.githubusercontent.com (release-assets.githubusercontent.com)|185.199.111.133|:443... connected.
HTTP request sent, awaiting response... 200 OK
Length: 314756836 (300M) [application/octet-stream]
Saving to: ‘FreeFEM-4.15-amd64-ubuntu24.04.deb’

FreeFEM-4.15-amd64-ubuntu24. 100%[=============================================>] 300.17M  6.44MB/s    in 50s

2026-02-07 17:52:37 (6.05 MB/s) - ‘FreeFEM-4.15-amd64-ubuntu24.04.deb’ saved [314756836/314756836]

gtheler@chalmers:~/psd$ sudo apt-get update && sudo apt-get install -y \
    freeglut3-dev liblapack-dev libhdf5-dev libgsl-dev \
    cmake libopenmpi-dev libtirpc-dev autoconf automake \
    libtool make python3 g++ gcc gfortran wget git flex \
    bison unzip patch pkg-config libxrender1 libxcursor1 \
    libxfixes3 libxft2 libxinerama1 libxrandr2 libxdamage1 \
    gmsh ./FreeFEM-4.15-amd64-ubuntu24.04.deb
[sudo] password for gtheler:
Hit:1 http://deb.debian.org/debian trixie InRelease
Hit:2 http://security.debian.org/debian-security trixie-security InRelease
Hit:3 http://deb.debian.org/debian trixie-updates InRelease
Hit:5 https://download.docker.com/linux/debian trixie InRelease
Hit:4 https://repository.spotify.com stable InRelease
Hit:6 https://dl.google.com/linux/chrome/deb stable InRelease
Reading package lists... Done
Reading package lists... Done
Building dependency tree... Done
Reading state information... Done
Note, selecting 'freefem' instead of './FreeFEM-4.15-amd64-ubuntu24.04.deb'
liblapack-dev is already the newest version (3.12.1-6).
liblapack-dev set to manually installed.
libhdf5-dev is already the newest version (1.14.5+repack-3).
libhdf5-dev set to manually installed.
libgsl-dev is already the newest version (2.8+dfsg-5).
cmake is already the newest version (3.31.6-2).
libopenmpi-dev is already the newest version (5.0.7-1).
libopenmpi-dev set to manually installed.
autoconf is already the newest version (2.72-3.1).
automake is already the newest version (1:1.17-4).
libtool is already the newest version (2.5.4-4).
libtool set to manually installed.
make is already the newest version (4.4.1-2).
python3 is already the newest version (3.13.5-1).
g++ is already the newest version (4:14.2.0-1).
gcc is already the newest version (4:14.2.0-1).
gfortran is already the newest version (4:14.2.0-1).
gfortran set to manually installed.
wget is already the newest version (1.25.0-2).
git is already the newest version (1:2.47.3-0+deb13u1).
unzip is already the newest version (6.0-29).
patch is already the newest version (2.8-2).
patch set to manually installed.
pkg-config is already the newest version (1.8.1-4).
pkg-config set to manually installed.
libxrender1 is already the newest version (1:0.9.12-1).
libxcursor1 is already the newest version (1:1.2.3-1).
libxfixes3 is already the newest version (1:6.0.0-2+b4).
libxft2 is already the newest version (2.3.6-1+b4).
libxft2 set to manually installed.
libxinerama1 is already the newest version (2:1.1.4-3+b4).
libxinerama1 set to manually installed.
libxrandr2 is already the newest version (2:1.5.4-1+b3).
libxdamage1 is already the newest version (1:1.1.6-1+b2).
gmsh is already the newest version (4.13.1+ds1-6).
gmsh set to manually installed.
The following additional packages will be installed:
  libfl-dev libfl2 libglut-dev libglut3.12
Suggested packages:
  bison-doc flex-doc
The following NEW packages will be installed:
  bison flex freefem freeglut3-dev libfl-dev libfl2 libglut-dev libglut3.12 libtirpc-dev
0 upgraded, 9 newly installed, 0 to remove and 162 not upgraded.
Need to get 2,309 kB/317 MB of archives.
After this operation, 6,756 kB of additional disk space will be used.
Get:1 /home/gtheler/psd/FreeFEM-4.15-amd64-ubuntu24.04.deb freefem amd64 4.15 [315 MB]
Get:2 http://deb.debian.org/debian trixie/main amd64 flex amd64 2.6.4-8.2+b4 [423 kB]
Get:3 http://deb.debian.org/debian trixie/main amd64 bison amd64 2:3.8.2+dfsg-1+b2 [1,170 kB]
Get:4 http://deb.debian.org/debian trixie/main amd64 libglut3.12 amd64 3.4.0-4 [139 kB]
Get:5 http://deb.debian.org/debian trixie/main amd64 libglut-dev amd64 3.4.0-4 [163 kB]
Get:6 http://deb.debian.org/debian trixie/main amd64 freeglut3-dev amd64 3.4.0-4 [53.4 kB]
Get:7 http://deb.debian.org/debian trixie/main amd64 libfl2 amd64 2.6.4-8.2+b4 [84.3 kB]
Get:8 http://deb.debian.org/debian trixie/main amd64 libfl-dev amd64 2.6.4-8.2+b4 [85.6 kB]
Get:9 http://deb.debian.org/debian trixie/main amd64 libtirpc-dev amd64 1.3.6+ds-1 [191 kB]
Fetched 2,309 kB in 2s (1,222 kB/s)
Selecting previously unselected package flex.
(Reading database ... 603359 files and directories currently installed.)
Preparing to unpack .../0-flex_2.6.4-8.2+b4_amd64.deb ...
Unpacking flex (2.6.4-8.2+b4) ...
Selecting previously unselected package bison.
Preparing to unpack .../1-bison_2%3a3.8.2+dfsg-1+b2_amd64.deb ...
Unpacking bison (2:3.8.2+dfsg-1+b2) ...
Selecting previously unselected package libglut3.12:amd64.
Preparing to unpack .../2-libglut3.12_3.4.0-4_amd64.deb ...
Unpacking libglut3.12:amd64 (3.4.0-4) ...
Selecting previously unselected package libglut-dev:amd64.
Preparing to unpack .../3-libglut-dev_3.4.0-4_amd64.deb ...
Unpacking libglut-dev:amd64 (3.4.0-4) ...
Selecting previously unselected package freeglut3-dev:amd64.
Preparing to unpack .../4-freeglut3-dev_3.4.0-4_amd64.deb ...
Unpacking freeglut3-dev:amd64 (3.4.0-4) ...
Selecting previously unselected package libfl2:amd64.
Preparing to unpack .../5-libfl2_2.6.4-8.2+b4_amd64.deb ...
Unpacking libfl2:amd64 (2.6.4-8.2+b4) ...
Selecting previously unselected package libfl-dev:amd64.
Preparing to unpack .../6-libfl-dev_2.6.4-8.2+b4_amd64.deb ...
Unpacking libfl-dev:amd64 (2.6.4-8.2+b4) ...
Selecting previously unselected package libtirpc-dev:amd64.
Preparing to unpack .../7-libtirpc-dev_1.3.6+ds-1_amd64.deb ...
Unpacking libtirpc-dev:amd64 (1.3.6+ds-1) ...
Selecting previously unselected package freefem.
Preparing to unpack .../8-FreeFEM-4.15-amd64-ubuntu24.04.deb ...
Unpacking freefem (4.15) ...
Setting up flex (2.6.4-8.2+b4) ...
Setting up libtirpc-dev:amd64 (1.3.6+ds-1) ...
Setting up libglut3.12:amd64 (3.4.0-4) ...
Setting up libfl2:amd64 (2.6.4-8.2+b4) ...
Setting up bison (2:3.8.2+dfsg-1+b2) ...
update-alternatives: using /usr/bin/bison.yacc to provide /usr/bin/yacc (yacc) in auto mode
Setting up libglut-dev:amd64 (3.4.0-4) ...
Setting up libfl-dev:amd64 (2.6.4-8.2+b4) ...
Setting up freeglut3-dev:amd64 (3.4.0-4) ...
Setting up freefem (4.15) ...
Processing triggers for man-db (2.13.1-1) ...
Processing triggers for install-info (7.1.1-1+b1) ...
Processing triggers for libc-bin (2.41-12) ...
gtheler@chalmers:~/psd$ autoreconf -i
configure.ac:289: installing './compile'
configure.ac:2: installing './install-sh'
configure.ac:2: installing './missing'
parallel-tests: installing './test-driver'
src/psd-preprocessor/Makefile.am: installing './depcomp'
gtheler@chalmers:~/psd$ ^[[200~./configure --prefix=${HOME}/psd~^C^C
gtheler@chalmers:~/psd$ ./configure --prefix=${HOME}/psd
checking for a BSD-compatible install... /usr/bin/install -c
checking whether sleep supports fractional seconds... yes
checking filesystem timestamp resolution... 0.01
checking whether build environment is sane... yes
checking for a race-free mkdir -p... /usr/bin/mkdir -p
checking for gawk... gawk
checking whether make sets $(MAKE)... yes
checking whether make supports nested variables... yes
checking xargs -n works... yes

--------------------------------------------------
         checking for operating system type
--------------------------------------------------

 OS identified: Linux
 Shared object sufix   : 'so'
checking for mpirun... mpirun
configure: WARNING:

 PSD can automatically build for you its dependencies
 FreeFEM, PETSc, MFront, MGIS, METIS,  ParMETIS, etc.
 A power user is adviced to build these on their own.
 In order to do so

 --with-dependencies=yes

 The wget command will be used to download these,  if
 you wish to provided already downloaded  files,  use
   --with-zipped_dependencies  option.

configure: WARNING:

 PSD can automatically build for you its dependencies
 FreeFEM, PETSc, MFront,  MGIS, METIS, ParMETIS, etc.
 from *.tar.gz files. In order to do so, provide the
 tar.gz files for dependencies in /ext folder of PSD

 --with-zipped_dependencies=yes

-------------------------------------------
    checking for a C++/MPIC++ compilers
-------------------------------------------
checking for g++... g++
checking whether the C++ compiler works... yes
checking for C++ compiler default output file name... a.out
checking for suffix of executables...
checking whether we are cross compiling... no
checking for suffix of object files... o
checking whether the compiler supports GNU C++... yes
checking whether g++ accepts -g... yes
checking for g++ option to enable C++11 features... none needed
checking whether make supports the include directive... yes (GNU style)
checking dependency style of g++... gcc3
checking for mpic++... mpic++
checking for MPI_Init... yes
checking for mpi.h... yes

---------------------------------------
    checking for a C/MPIC compilers
---------------------------------------
checking for gcc... gcc
checking whether the compiler supports GNU C... yes
checking whether gcc accepts -g... yes
checking for gcc option to enable C11 features... none needed
checking whether gcc understands -c and -o together... yes
checking dependency style of gcc... gcc3
checking for mpicc... mpicc
checking for mpi.h... yes

----------------------------------------------
    checking for a Fortran/MPIFC compilers
----------------------------------------------
checking for gfortran... gfortran
checking whether the compiler supports GNU Fortran... yes
checking whether gfortran accepts -g... yes
checking for mpif90... mpif90
checking for mpif.h... yes

--------------------------------------------------
    checking for a Fortran-77/MPIF77 compilers
--------------------------------------------------
checking for g77... no
checking for xlf... no
checking for f77... f77
checking whether the compiler supports GNU Fortran 77... yes
checking whether f77 accepts -g... yes
checking for mpif77... mpif77
checking for mpif.h... yes

----------------------------------------------
         checking for a Gmsh mesher
----------------------------------------------
checking for gmsh... gmsh

Using internal dependencies... no
checking for Gmsh...           gmsh
Gmsh support...

-------------------------------------------------
         checking for a FreeFEM compilers
-------------------------------------------------
checking for FreeFem++... FreeFem++
checking for ff-c++... ff-c++
checking for ff-get-dep... ff-get-dep
checking for ff-mpirun... ff-mpirun

Using internal dependencies... no
checking for FreeFem++...      FreeFem++
checking for ff-c++...         ff-c++
checking for ff-get-dep...     ff-get-dep
checking for ff-mpirun...      ff-mpirun
FreeFEM support...             yes

---------------------------------------------------
         checking for a MFront/MGIS support
---------------------------------------------------
checking for mfront... no
configure: Using mfront:
checking for mfront... no
checking for MGIS library directory... not found

MGIS Prefix:
MGIS Support:      no
MFront Prefix:
MFront support:    yes

----------------------------------------------
         checking for a salome support
----------------------------------------------
configure: WARNING:
 PSD will not be configured with salome,
 you did not provide --with-salome flag so

 SALOME SUPPORT = OFF

 To have salome support in PSD, please correctly
 provide both  --with-salome flag.

-------------------------------------------
         checking for a med support
-------------------------------------------

Using internal dependencies... no
checking for hdf5...          no
checking for medfile...       no
checking for medcoupling...   no
MED support...                no

configure: WARNING:
 top-ii-vol will not be configured with MED support.
 You must provide either:

  --with-dependencies
    OR
  --with-hdf5
  --with-medfile
  --with-medcoupling

 MED SUPPORT = OFF

----------------------------------------------
       generating the required makefiles
----------------------------------------------
checking that generated files are newer than configure... done
configure: creating ./config.status
config.status: creating Makefile
config.status: creating data/Makefile
config.status: creating data/meshes/2D/Geo-Files/Makefile
config.status: creating data/meshes/3D/Geo-Files/Makefile
config.status: creating ext/Makefile
config.status: creating src/Makefile
config.status: creating src/plugins/Makefile
config.status: creating src/ff-files/Makefile
config.status: creating src/plugins/go-fast-plugins/Makefile
config.status: creating src/plugins/go-fast-plugins/cpp/Makefile
config.status: creating src/plugins/go-fast-plugins/test/Makefile
config.status: creating src/plugins/medio/Makefile
config.status: creating src/plugins/medio/cpp/Makefile
config.status: creating src/plugins/medio/test/Makefile
config.status: creating src/plugins/mfront/Makefile
config.status: creating src/plugins/mfront/cpp/Makefile
config.status: creating src/plugins/mfront/law/Makefile
config.status: creating src/plugins/mfront/law/test/Makefile
config.status: creating src/plugins/mfront/test/Makefile
config.status: creating src/plugins/soil-laws/Makefile
config.status: creating src/plugins/soil-laws/cpp/Makefile
config.status: creating src/plugins/soil-laws/hujeux-law-lib/Makefile
config.status: creating src/plugins/soil-laws/test/Makefile
config.status: creating src/plugins/sparsity-visualizer/Makefile
config.status: creating src/plugins/sparsity-visualizer/cpp/Makefile
config.status: creating src/plugins/sparsity-visualizer/test/Makefile
config.status: creating src/plugins/top-ii-vol/Makefile
config.status: creating src/plugins/top-ii-vol/cpp/Makefile
config.status: creating src/plugins/top-ii-vol/test/Makefile
config.status: creating src/psd-preprocessor/Makefile
config.status: creating src/psd-solve/Makefile
config.status: creating test/Makefile
config.status: creating test/elastodynamics/Makefile
config.status: creating test/elasto-plastic/Makefile
config.status: creating test/fracture-mechanics/Makefile
config.status: creating test/linear-elasticity/Makefile
config.status: creating test/soildynamics/Makefile
config.status: creating test/poisson/Makefile
config.status: creating config.h
config.status: executing depfiles commands

------------------------------------------
                  status
------------------------------------------

 PSD version 2.6 configured successfully.

 PSD install  path  : '/home/gtheler/psd'
 PSD tutoirlas path : '/home/gtheler/PSD-tutorials'

 C++ compiler       : 'g++'
 C compiler         : 'gcc'
 FC compiler        : 'gfortran'
 F77 compiler       : 'f77'

 C/C++ compiler flags   : '-g -O2'
 F77/FC compiler flags  : '-g -O2'

 MPIRUN             : 'mpirun'
 MPIC++ compiler    : 'mpic++'
 MPIC   compiler    : 'mpicc'
 MPIF77 compiler    : 'mpif77'
 MPIFC  compiler    : 'mpif90'

 FreeFEM compiler   : 'FreeFem++'
 ff-c++  compiler   : 'ff-c++'
 ff-mpi  compiler   : 'ff-mpirun'

 Mfront      : YES
 Gmsh        : YES
 Hdf5        : NO
 medfile     : NO
 medcoupling : NO
 med support : NO
 salome      : NO
 PSD Seq     : YES
 PSD Par     : YES

gtheler@chalmers:~/psd$ make
make  all-recursive
make[1]: Entering directory '/home/gtheler/psd'
Making all in ext
make[2]: Entering directory '/home/gtheler/psd/ext'

make[2]: Leaving directory '/home/gtheler/psd/ext'
Making all in src
make[2]: Entering directory '/home/gtheler/psd/src'
Making all in plugins
make[3]: Entering directory '/home/gtheler/psd/src/plugins'
Making all in top-ii-vol
make[4]: Entering directory '/home/gtheler/psd/src/plugins/top-ii-vol'
Making all in cpp
make[5]: Entering directory '/home/gtheler/psd/src/plugins/top-ii-vol/cpp'
ff-c++ -auto -mpi top-ii-vol.cpp
/usr/local/bin/ff-c++: 13: -rdynamic: not found
eval /usr/local/bin/ff-c++ '-mpi' top-ii-vol.cpp
/usr/local/bin/ff-c++: 13: -rdynamic: not found
/usr/bin/mpic++ -c -fPIC '-I/usr/local/lib/ff++/4.15/include' -g -O2 -DNDEBUG -O3 -mtune=generic -std=gnu++14 -DBAMG_LONG_LONG -DNCHECKPTR -fPIC -I/usr/local/ff-petsc/r/include/suitesparse -I/usr/local/ff-petsc/r/include -I/usr/include -I/usr/include/hdf5/serial 'top-ii-vol.cpp'

/usr/bin/mpic++ -shared -fPIC -g -O2 -DNDEBUG -O3 -mtune=generic -std=gnu++14 -DBAMG_LONG_LONG -DNCHECKPTR -fPIC -I/usr/local/ff-petsc/r/include/suitesparse -I/usr/local/ff-petsc/r/include -I/usr/include -I/usr/include/hdf5/serial 'top-ii-vol.o' -o top-ii-vol.so -L/usr/lib/x86_64-linux-gnu/openmpi/lib -lmpi_cxx -lmpi
/usr/bin/ld: cannot find -lmpi_cxx: No such file or directory
collect2: error: ld returned 1 exit status
make[5]: *** [Makefile:448: top-ii-vol.so] Error 1
make[5]: Leaving directory '/home/gtheler/psd/src/plugins/top-ii-vol/cpp'
make[4]: *** [Makefile:348: all-recursive] Error 1
make[4]: Leaving directory '/home/gtheler/psd/src/plugins/top-ii-vol'
make[3]: *** [Makefile:348: all-recursive] Error 1
make[3]: Leaving directory '/home/gtheler/psd/src/plugins'
make[2]: *** [Makefile:358: all-recursive] Error 1
make[2]: Leaving directory '/home/gtheler/psd/src'
make[1]: *** [Makefile:414: all-recursive] Error 1
make[1]: Leaving directory '/home/gtheler/psd'
[…]
