### 2026-02-07 mohd-afeef-badri/psd#56 @gtheler
I started all over again from a fresh laptop with Debian Trixie without any luck.
Again the complain of missing `libmpi_cxx.so` but this time I could not resolve with symbolic links. Also there was a missing `libhdf5_serial.so` which I am not sure where it must come from since I have all the hd5 flavors.

I then checked and the FreeFEM version in Debian Trixie is 4.15, so I installed it but it is not recognized but configure:

```
gtheler@chalmers:~$ sudo apt-cache search freefem
freefem - PDE oriented language using Finite Element Method
freefem-doc - Documentation for FreeFEM (html and pdf)
freefem-examples - Example files for FreeFEM
libfreefem-dev - Development library, header files and manpages
libfreefem-doc - Documentation for FreeFEM development
libfreefem0 - Shared libraries for FreeFEM
freefem++ - Provides the binaries of the FreeFem++ FE suite
freefem++-doc - Provides the documentation of the FreeFem++ FE suite
libfreefem++ - Provides the shared libraries of the FreeFem++ FE suite
libfreefem++-dev - Provides the development files of the FreeFem++ FE suite
gtheler@chalmers:~$ sudo apt install libfreefem++-dev
Installing:
  libfreefem++-dev

Installing dependencies:
  libfreefem++  libmumps-seq-5.7  libtet1.5

Summary:
  Upgrading: 0, Installing: 4, Removing: 0, Not Upgrading: 162
  Download size: 4,549 kB
  Space needed: 33.5 MB / 105 GB available

Continue? [Y/n]
Get:1 http://deb.debian.org/debian trixie/main amd64 libmumps-seq-5.7 amd64 5.7.3-3+b1 [2,006 kB]
Get:2 http://deb.debian.org/debian trixie/main amd64 libtet1.5 amd64 1.5.0-5+b2 [239 kB]
Get:3 http://deb.debian.org/debian trixie/main amd64 libfreefem++ amd64 4.15+dfsg-1 [1,966 kB]
Get:4 http://deb.debian.org/debian trixie/main amd64 libfreefem++-dev amd64 4.15+dfsg-1 [338 kB]
Fetched 4,549 kB in 2s (2,821 kB/s)
Selecting previously unselected package libmumps-seq-5.7:amd64.
(Reading database ... 603812 files and directories currently installed.)
Preparing to unpack .../libmumps-seq-5.7_5.7.3-3+b1_amd64.deb ...
Unpacking libmumps-seq-5.7:amd64 (5.7.3-3+b1) ...
Selecting previously unselected package libtet1.5:amd64.
Preparing to unpack .../libtet1.5_1.5.0-5+b2_amd64.deb ...
Unpacking libtet1.5:amd64 (1.5.0-5+b2) ...
Selecting previously unselected package libfreefem++.
Preparing to unpack .../libfreefem++_4.15+dfsg-1_amd64.deb ...
Unpacking libfreefem++ (4.15+dfsg-1) ...
Selecting previously unselected package libfreefem++-dev.
Preparing to unpack .../libfreefem++-dev_4.15+dfsg-1_amd64.deb ...
Unpacking libfreefem++-dev (4.15+dfsg-1) ...
Setting up libmumps-seq-5.7:amd64 (5.7.3-3+b1) ...
Setting up libtet1.5:amd64 (1.5.0-5+b2) ...
Setting up libfreefem++ (4.15+dfsg-1) ...
Setting up libfreefem++-dev (4.15+dfsg-1) ...
Processing triggers for libc-bin (2.41-12) ...
gtheler@chalmers:~$ git clone git@github.com:mohd-afeef-badri/psd.git
Cloning into 'psd'...
remote: Enumerating objects: 10200, done.
remote: Total 10200 (delta 159), reused 137 (delta 123), pack-reused 9950 (from 2)
Resolving deltas: 100% (6782/6782), done.
gtheler@chalmers:~$ cd psd/
gtheler@chalmers:~/psd$ autoreconf -i
configure.ac:289: installing './compile'
configure.ac:2: installing './install-sh'
configure.ac:2: installing './missing'
parallel-tests: installing './test-driver'
src/psd-preprocessor/Makefile.am: installing './depcomp'
gtheler@chalmers:~/psd$ ./configure --prefix=${HOME}/psd
checking for a BSD-compatible install... /usr/bin/install -c
checking whether sleep supports fractional seconds... yes
checking filesystem timestamp resolution... 0.01
checking whether build environment is sane... yes
checking for a race-free mkdir -p... /usr/bin/mkdir -p
checking for gawk... gawk
checking whether make sets $(MAKE)... yes
checking whether make supports nested variables... yes
checking xargs -n works... yes

--------------------------------------------------
         checking for operating system type
--------------------------------------------------

 OS identified: Linux
 Shared object sufix   : 'so'
checking for mpirun... mpirun
configure: WARNING:

 PSD can automatically build for you its dependencies
 FreeFEM, PETSc, MFront, MGIS, METIS,  ParMETIS, etc.
 A power user is adviced to build these on their own.
 In order to do so

 --with-dependencies=yes

 The wget command will be used to download these,  if
 you wish to provided already downloaded  files,  use
   --with-zipped_dependencies  option.

configure: WARNING:

 PSD can automatically build for you its dependencies
 FreeFEM, PETSc, MFront,  MGIS, METIS, ParMETIS, etc.
 from *.tar.gz files. In order to do so, provide the
 tar.gz files for dependencies in /ext folder of PSD

 --with-zipped_dependencies=yes

-------------------------------------------
    checking for a C++/MPIC++ compilers
-------------------------------------------
checking for g++... g++
checking whether the C++ compiler works... yes
checking for C++ compiler default output file name... a.out
checking for suffix of executables...
checking whether we are cross compiling... no
checking for suffix of object files... o
checking whether the compiler supports GNU C++... yes
checking whether g++ accepts -g... yes
checking for g++ option to enable C++11 features... none needed
checking whether make supports the include directive... yes (GNU style)
checking dependency style of g++... gcc3
checking for mpic++... mpic++
checking for MPI_Init... yes
checking for mpi.h... yes

---------------------------------------
    checking for a C/MPIC compilers
---------------------------------------
checking for gcc... gcc
checking whether the compiler supports GNU C... yes
checking whether gcc accepts -g... yes
checking for gcc option to enable C11 features... none needed
checking whether gcc understands -c and -o together... yes
checking dependency style of gcc... gcc3
checking for mpicc... mpicc
checking for mpi.h... yes

----------------------------------------------
    checking for a Fortran/MPIFC compilers
----------------------------------------------
checking for gfortran... gfortran
checking whether the compiler supports GNU Fortran... yes
checking whether gfortran accepts -g... yes
checking for mpif90... mpif90
checking for mpif.h... yes

--------------------------------------------------
    checking for a Fortran-77/MPIF77 compilers
--------------------------------------------------
checking for g77... no
checking for xlf... no
checking for f77... f77
checking whether the compiler supports GNU Fortran 77... yes
checking whether f77 accepts -g... yes
checking for mpif77... mpif77
checking for mpif.h... yes

----------------------------------------------
         checking for a Gmsh mesher
----------------------------------------------
checking for gmsh... gmsh

Using internal dependencies... no
checking for Gmsh...           gmsh
Gmsh support...

-------------------------------------------------
         checking for a FreeFEM compilers
-------------------------------------------------
checking for FreeFem++... no
checking for ff-c++... no
checking for ff-get-dep... no
checking for ff-mpirun... no

Using internal dependencies... no
checking for FreeFem++...      no
checking for ff-c++...         no
checking for ff-get-dep...     no
checking for ff-mpirun...      no
FreeFEM support...             no

configure: error:
 Required  program  'FreeFem++'  not  found.
 You must provide either:

  --with-dependencies
    OR
  --with-FreeFEM=your/FreeFEM/path/bin

 FREEFEM SUPPORT = OFF

gtheler@chalmers:~/psd$
```
