### 2026-02-15 #9411 @mohd-afeef-badri
@nrichart thanks for the review, the results provided in the paper were large scale runs on HPC machines (the figures are from peer-reviewed published papers), moreover the earthquake simulation was with specific data that is not shareable. However:

- Scaled down eathquake case with a similar mesh (low resolution) and with arbitary prperties is provided as a tutorial case https://mohd-afeef-badri.github.io/psd/#/soildynamics . These are desigened to run on multi CPU PC setup, so you can test them out. You can see tutorial 2, 3, 4.

- Same for fracture mechanics there are some tutorials https://mohd-afeef-badri.github.io/psd/#/fracture-mechanics.

Please let me know these are ok for you
