### 2026-02-15 #9411 @nrichart
@mohd-afeef-badri I understand the size of this simulations I did run the tutorials, but they are quite small. In your paper and in the documentation you say that the code is targeting HPC machines with billions of degree of freedom.
I wanted to run such a case, I do not need to run the full simulation time but at least a few time steps.

Is there a way to refine the mesh of the tutorials to run on more than 4 cores. I wanted to test at least a few hundred cores.
