### 2026-02-15 mohd-afeef-badri/psd#65 @nrichart
[JOSS Review] Community guidelines
I think the readme is missing a section for the **Community guidelines**

In the checklists in https://github.com/openjournals/joss-reviews/issues/9411 you can read this:
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

@mohd-afeef-badri  you just have a section asking to contact you. Is this the preferred way for issues and third-party contributions ?
