### 2026-02-15 32816d3
Create CODE_OF_CONDUCT.md
