### 2026-02-15 mohd-afeef-badri/psd#65 @mohd-afeef-badri
@nrichart Indeed this was missing. in #66  I have added the guidelines and code of conduct, further I have referenced these in the Readme. Is that ok for you ?
