### 2026-02-15 mohd-afeef-badri/psd#56 @mohd-afeef-badri
@gtheler I have finally managed to write a spack package for PSD.  https://github.com/mohd-afeef-badri/spack_psd
You will find the Readme there that helps install PSD via spack. I tested it on my machine UB24.04 and all seems to work. Can you please try with this method to see if everything is ok. If this procedure works I will include this in the offical documenatation as a way to install PSD. Thank you again for your patience.
