### 2026-02-16 #9411 @mohd-afeef-badri
> [@mohd-afeef-badri](https://github.com/mohd-afeef-badri) I understand the size of this simulations I did run the tutorials, but they are quite small. In your paper and in the documentation you say that the code is targeting HPC machines with billions of degree of freedom. I wanted to run such a case, I do not need to run the full simulation time but at least a few time steps.
>
> Is there a way to refine the mesh of the tutorials to run on more than 4 cores. I wanted to test at least a few hundred cores.

Thank you for the comment, i somehow missed responding to this yesterday.

Yes, the tutorial meshes can be enlarged directly using the built-in uniform refinement option `-split n` this is mentioned in the tutorials [here](https://mohd-afeef-badri.github.io/psd/#/elastodynamics?id=exercise-2-mesh-refinement-and-convergence-study) , [and here](https://mohd-afeef-badri.github.io/psd/#/linear-elasticity?id=exercise-1-mesh-refinement-and-convergence-study) in the tutorials. Please note this flag needs to be provided while solving step with `PSD_Solve`.

 The option `-split n` applies `n-1` successive uniform refinement steps to the input mesh before the simulation starts. At each split, every element is subdivided (e.g. in 2D each element is split into 4, and in 3D into 8), so the total number of cells and therefore the number of degrees of freedom grows exponentially ($4^{n-1}$ $8^{n-1}$)with the number of splits.

In practice, this allows any of the `PSD` cases to be turned into large-scale benchmarks without modifying the input files, and there is an advantage to this because splits are performed post mesh partitioning.

Please let me know if this is ok for you.
