### 2026-02-16 mohd-afeef-badri/psd#56 @gtheler
```
gtheler@tom:~$ git clone --depth=2 https://github.com/spack/spack.git
Cloning into 'spack'...
remote: Enumerating objects: 2369, done.
remote: Total 2369 (delta 314), reused 1302 (delta 236), pack-reused 0 (from 0)
Resolving deltas: 100% (314/314), done.
gtheler@tom:~$ cd spack/
gtheler@tom:~/spack$ cd share/
gtheler@tom:~/spack/share/spack$ source setup-env.sh
gtheler@tom:~/spack/share/spack$ cd ..
gtheler@tom:~/spack/share$ cd ..
gtheler@tom:~/spack$ cd ..
gtheler@tom:~$ git clone https://github.com/mohd-afeef-badri/spack_psd
Cloning into 'spack_psd'...
remote: Enumerating objects: 23, done.
remote: Total 23 (delta 1), reused 20 (delta 1), pack-reused 0 (from 0)
Receiving objects: 100% (23/23), 8.58 KiB | 8.58 MiB/s, done.
Resolving deltas: 100% (1/1), done.
gtheler@tom:~$ spack repo add spack_psd/spack_repo/psd/
==> Added repo to config with name 'psd'.
gtheler@tom:~$ spack install psd
remote: Enumerating objects: 19933, done.
remote: Total 19933 (delta 1346), reused 14085 (delta 1172), pack-reused 0 (from 0)
==> Compilers have been configured automatically from PATH inspection
==> Error: Malformed string https://mirror.spack.io/bootstrap/github-actions/v0.1
gtheler@tom:~$
```
