### 2026-02-16 mohd-afeef-badri/psd#56 @mohd-afeef-badri
@gtheler I tried to replicate your issue but on my side everything works ok

```
~ $which spack
~ $spack --version
Command 'spack' not found, did you mean:
  command 'slack' from snap slack (4.46.99)
  command 'cpack' from snap cmake (4.2.3)
  command 'fpack' from deb libcfitsio-bin (4.3.0-2)
  command 'pack' from deb liballegro4-dev (2:4.4.3.1-4)
  command 'spock' from deb python3-sardana (3.0.3-1)
  command 'slack' from deb slack (1:0.15.2-11)
  command 'stack' from deb haskell-stack (2.9.3.1-1)
  command 'hpack' from deb hpack (0.35.2-1)
  command 'cpack' from deb cmake (3.27.8-1build1)
  command 'apack' from deb atool (0.39.0-13)
  command 'ipack' from deb exult-studio (1.8-2)
  command 'mpack' from deb mpack (1.6-18)
See 'snap info <snapname>' for additional versions.
~ $git clone --depth=2 https://github.com/spack/spack.git
Cloning into 'spack'...
remote: Enumerating objects: 2375, done.
remote: Total 2375 (delta 320), reused 1307 (delta 237), pack-reused 0 (from 0)
Resolving deltas: 100% (320/320), done.
~ $cd spack/
spack $cd share/
share $cd spack/
spack $source setup-env.sh
spack $cd ..
share $cd ..
spack $cd ..
~ $git clone https://github.com/mohd-afeef-badri/spack_psd
Cloning into 'spack_psd'...
remote: Enumerating objects: 29, done.
remote: Total 29 (delta 1), reused 20 (delta 1), pack-reused 0 (from 0)
Receiving objects: 100% (29/29), 9.78 KiB | 9.78 MiB/s, done.
Resolving deltas: 100% (1/1), done.
~ $spack repo add spack_psd/spack_repo/psd/
==> Added repo to config with name 'psd'.
~ $spack install psd
[+] /usr (external gmake-4.3-sepxc6faoksxxurr2feyie3drmmatszu)
....
....
==> Installing psd-main-dr7xd4mw74o6ogpeipcf756iizhq6qwz [61/61]
Initialized empty Git repository in /tmp/tmpeu6l170i/psd/.git/
==> No patches needed for psd
==> psd: Executing phase: 'autoreconf'
==> psd: Executing phase: 'configure'
==> psd: Executing phase: 'build'
==> psd: Executing phase: 'install'
==> psd: Successfully installed psd-main-dr7xd4mw74o6ogpeipcf756iizhq6qwz
  Stage: 8.78s.  Autoreconf: 2.47s.  Configure: 1.67s.  Build: 53.03s.  Install: 0.27s.  Post-install: 0.01s.  Total: 1m 6.35s
[+] /home/afeef/spack/opt/spack/linux-skylake/psd-main-dr7xd4mw74o6ogpeipcf756iizhq6qwz
~ $

```

I guess the issue you are facing is of `spack` not PSD, maybe you have old spack config there on your PC.

Can you try to check `spack bootstrap list` ? `spack --version` to check if you using the correct version. also if you maybe remove  `rm  -r f ~/.spack` before that would maybe help.

Also what is strange your spack github-actions for bootstrap is v0.1 while for me it is v0.6 as is the case here https://spack.readthedocs.io/en/latest/bootstrapping.html . This makes me doubt that there might be old spack config. Hope this helps.
