### 2026-02-17 mohd-afeef-badri/psd#56 @gtheler
Indeed, apparently I had a stale `.spack` directory, God knows since when.
I made some progress but after a couple of hours I got an error

```
==> No binary for parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m found: installing from source
==> Installing parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m [64/68]
==> Using cached archive: /home/gtheler/spack/var/spack/cache/_source-cache/archive/0b/0baec7914e49a26bdbb849ab64dcd92147eff79ac02ef3b2599cb05104901a7a.tar.gz
==> Applied patch /home/gtheler/.spack/package_repos/fncqgg4/repos/spack_repo/builtin/packages/parmmg/parmmg_cmake_patch.diff
==> parmmg: Executing phase: 'cmake'
==> parmmg: Executing phase: 'build'
==> Error: ProcessError: Command exited with status 2:
    '/home/gtheler/spack/opt/spack/linux-skylake/gmake-4.4.1-d5otqo54o7qa4i2eg7ixnlxkmvimiszo/bin/make' '-j8'

9 errors found in build log:
     434    /home/gtheler/spack/opt/spack/linux-skylake/gmake-4.4.1-d5otqo54o7qa4i2eg7ixnlxkmvimiszo/bin/make  -f CMakeFiles/parmmg.
            dir/build.make CMakeFiles/parmmg.dir/build
     435    make[2]: Entering directory '/tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m/spack-bu
            ild-zilbpzd'
     436    [ 98%] Building C object CMakeFiles/parmmg.dir/src/parmmg.c.o
     437    /home/gtheler/spack/opt/spack/linux-skylake/compiler-wrapper-1.0-udjq6iooiel5itphzuxzgerik5gqqopp/libexec/spack/gcc/gcc
            -DNDEBUG -DUSE_SCOTCH -I/home/gtheler/spack/opt/spack/linux-skylake/openmpi-5.0.9-habcjpoz5sosphjxageq4eubw427pgks/inclu
            de -I/usr/include/hdf5/openmpi -I/home/gtheler/spack/opt/spack/linux-skylake/mmg-5.8.0-y46dpaq4lfqxvaw3acv7vgihlg2w2opd/
            include -I/home/gtheler/spack/opt/spack/linux-skylake/mmg-5.8.0-y46dpaq4lfqxvaw3acv7vgihlg2w2opd/include/mmg/mmg3d -I/ho
            me/gtheler/spack/opt/spack/linux-skylake/mmg-5.8.0-y46dpaq4lfqxvaw3acv7vgihlg2w2opd/include/mmg/common -I/home/gtheler/s
            pack/opt/spack/linux-skylake/metis-5.1.0-fhz7vhqrcoeggumdu5xjm4sd4qky5zul/include -I/tmp/gtheler/spack-stage/spack-stage
            -parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m/spack-src/src -I/tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzd
            yoe7myqgbf54jyzipfkrms54m/spack-build-zilbpzd/include -I/tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzdyoe7myq
            gbf54jyzipfkrms54m/spack-build-zilbpzd/src/parmmg -DUSE_METIS -DUSE_MPI  -DUSE_HDF5 -O3 -DNDEBUG -std=gnu99 -fPIE -MD -M
            T CMakeFiles/parmmg.dir/src/parmmg.c.o -MF CMakeFiles/parmmg.dir/src/parmmg.c.o.d -o CMakeFiles/parmmg.dir/src/parmmg.c.
            o -c /tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m/spack-src/src/parmmg.c
     438    [100%] Linking C executable bin/parmmg_O3
     439    /home/gtheler/spack/opt/spack/linux-skylake/cmake-3.31.11-iujypaeonsd5hwj32gphyh42lxtrpfc3/bin/cmake -E cmake_link_scrip
            t CMakeFiles/parmmg.dir/link.txt --verbose=1
  >> 440    /usr/bin/ld: /home/gtheler/spack/opt/spack/linux-skylake/openssl-3.6.1-piwzi4skr3udc5nvwtbf5dw3izima5u7/lib64/libssl.so.
            3: undefined reference to `CRYPTO_malloc_array@@OPENSSL_3.6.0'
  >> 441    /usr/bin/ld: /home/gtheler/spack/opt/spack/linux-skylake/openssl-3.6.1-piwzi4skr3udc5nvwtbf5dw3izima5u7/lib64/libssl.so.
            3: undefined reference to `OPENSSL_sk_set_thunks@@OPENSSL_3.6.0'
  >> 442    /usr/bin/ld: /home/gtheler/spack/opt/spack/linux-skylake/openssl-3.6.1-piwzi4skr3udc5nvwtbf5dw3izima5u7/lib64/libssl.so.
            3: undefined reference to `CRYPTO_calloc@@OPENSSL_3.6.0'
  >> 443    /usr/bin/ld: /home/gtheler/spack/opt/spack/linux-skylake/openssl-3.6.1-piwzi4skr3udc5nvwtbf5dw3izima5u7/lib64/libssl.so.
            3: undefined reference to `CRYPTO_realloc_array@@OPENSSL_3.6.0'
  >> 444    /usr/bin/ld: /home/gtheler/spack/opt/spack/linux-skylake/openssl-3.6.1-piwzi4skr3udc5nvwtbf5dw3izima5u7/lib64/libssl.so.
            3: undefined reference to `X509_STORE_CTX_set_ocsp_resp@@OPENSSL_3.6.0'
  >> 445    collect2: error: ld returned 1 exit status
     446    /home/gtheler/spack/opt/spack/linux-skylake/compiler-wrapper-1.0-udjq6iooiel5itphzuxzgerik5gqqopp/libexec/spack/gcc/gcc
            -DUSE_METIS -DUSE_MPI  -DUSE_HDF5 -O3 -DNDEBUG  -Wl,-rpath -Wl,/home/gtheler/spack/opt/spack/linux-skylake/openmpi-5.0.9
            -habcjpoz5sosphjxageq4eubw427pgks/lib -Wl,--dependency-file=CMakeFiles/parmmg.dir/link.d CMakeFiles/parmmg.dir/src/parmm
            g.c.o -o bin/parmmg_O3  -Wl,-rpath,/tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m/sp
            ack-build-zilbpzd/lib:/home/gtheler/spack/opt/spack/linux-skylake/metis-5.1.0-fhz7vhqrcoeggumdu5xjm4sd4qky5zul/lib:/home
            /gtheler/spack/opt/spack/linux-skylake/mmg-5.8.0-y46dpaq4lfqxvaw3acv7vgihlg2w2opd/lib:/home/gtheler/spack/opt/spack/linu
            x-skylake/scotch-7.0.10-7afou4a4d2o6drhbti4iqlkg44y3wvsp/lib:/home/gtheler/spack/opt/spack/linux-skylake/openmpi-5.0.9-h
            abcjpoz5sosphjxageq4eubw427pgks/lib:/usr/lib/x86_64-linux-gnu/hdf5/openmpi:/home/gtheler/spack/opt/spack/linux-skylake/z
            lib-ng-2.3.2-rast3fxi3d5qbyqewaoa72ernu7db3u3/lib:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
            :::::::::::: lib/libparmmg.so.1.5.0 /home/gtheler/spack/opt/spack/linux-skylake/metis-5.1.0-fhz7vhqrcoeggumdu5xjm4sd4qky
            5zul/lib/libmetis.so /home/gtheler/spack/opt/spack/linux-skylake/mmg-5.8.0-y46dpaq4lfqxvaw3acv7vgihlg2w2opd/lib/libmmg.s
            o /home/gtheler/spack/opt/spack/linux-skylake/scotch-7.0.10-7afou4a4d2o6drhbti4iqlkg44y3wvsp/lib/libscotch.so /home/gthe
            ler/spack/opt/spack/linux-skylake/scotch-7.0.10-7afou4a4d2o6drhbti4iqlkg44y3wvsp/lib/libscotcherrexit.so -lz -lm -lrt /h
            ome/gtheler/spack/opt/spack/linux-skylake/openmpi-5.0.9-habcjpoz5sosphjxageq4eubw427pgks/lib/libmpi.so /home/gtheler/spa
            ck/opt/spack/linux-skylake/openmpi-5.0.9-habcjpoz5sosphjxageq4eubw427pgks/lib/libmpi.so /usr/lib/x86_64-linux-gnu/hdf5/o
            penmpi/libhdf5.so /usr/lib/x86_64-linux-gnu/libcrypto.so /usr/lib/x86_64-linux-gnu/libcurl.so /usr/lib/x86_64-linux-gnu/
            libsz.so /home/gtheler/spack/opt/spack/linux-skylake/zlib-ng-2.3.2-rast3fxi3d5qbyqewaoa72ernu7db3u3/lib/libz.so /usr/lib
            /x86_64-linux-gnu/libdl.a -lm -lm
  >> 447    make[2]: *** [CMakeFiles/parmmg.dir/build.make:119: bin/parmmg_O3] Error 1
     448    make[2]: Leaving directory '/tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m/spack-bui
            ld-zilbpzd'
  >> 449    make[1]: *** [CMakeFiles/Makefile2:596: CMakeFiles/parmmg.dir/all] Error 2
     450    make[1]: Leaving directory '/tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m/spack-bui
            ld-zilbpzd'
  >> 451    make: *** [Makefile:139: all] Error 2

See build log for details:
  /tmp/gtheler/spack-stage/spack-stage-parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m/spack-build-out.txt

==> Warning: Skipping build of petsc-3.24.4-qcchs66vajb5wtfu2lzixhabwb6hh6tx since parmmg-1.5.0-zilbpzdyoe7myqgbf54jyzipfkrms54m failed
==> Warning: Skipping build of slepc-3.24.2-lax63arj4m3s624hth4dplc7lzdxnbxw since petsc-3.24.4-qcchs66vajb5wtfu2lzixhabwb6hh6tx failed
==> Warning: Skipping build of freefempsd-4.15-izueiwktm6wirohql47wvte2lgomtfdk since slepc-3.24.2-lax63arj4m3s624hth4dplc7lzdxnbxw failed
==> Warning: Skipping build of psd-main-fvlqz5vddaivpkh77vmtgxmqv44ou6iw since freefempsd-4.15-izueiwktm6wirohql47wvte2lgomtfdk failed
==> Error: psd-main-fvlqz5vddaivpkh77vmtgxmqv44ou6iw: Package was not installed
==> Error: Installation request failed.  Refer to reported errors for failing package(s).
```

Not sure why a FEM library needs crypto stuff.
