### 2026-02-17 mohd-afeef-badri/psd#56 @mohd-afeef-badri
@gtheler Thanks for trying it again. Indeed this is some progress.

It seems  `Parmmg`  has a unstable recipe for your OS.

Since `Parmmg` and `mmg` are not essential dependencies for PSD (needed for experimental mesh adaption), I have removed these from spack recipe for PSD (https://github.com/mohd-afeef-badri/spack_psd/commit/85bbe0789dc4e32e726776e629d90ccbc37511c5).

Can you please git pull `spack_psd` and retry please.
