### 2026-02-19 mohd-afeef-badri/psd#56 @gtheler
Nailed it (almost)!

```
==> No binary for psd-main-uulo4hjj6ebwydr7jtil4ahblx4jvymr found: installing from source
==> Installing psd-main-uulo4hjj6ebwydr7jtil4ahblx4jvymr [61/61]
Initialized empty Git repository in /tmp/tmpuag2lr6d/psd/.git/
==> No patches needed for psd
==> psd: Executing phase: 'autoreconf'
==> psd: Executing phase: 'configure'
==> psd: Executing phase: 'build'
==> psd: Executing phase: 'install'
==> psd: Successfully installed psd-main-uulo4hjj6ebwydr7jtil4ahblx4jvymr
  Stage: 8.31s.  Autoreconf: 2.80s.  Configure: 2.32s.  Build: 50.65s.  Install: 0.46s.  Post-install: 0.02s.  Total: 1m 4.74s
[+] /home/gtheler/spack/opt/spack/linux-skylake/psd-main-uulo4hjj6ebwydr7jtil4ahblx4jvymr
```

Continuing with the README
```
gtheler@tom:~/spack_psd$ spack load psd
gtheler@tom:~/spack_psd$ spack find psd
-- linux-debian13-skylake / %cxx=gcc@14.2.0 ---------------------
psd@main
==> 1 installed package
gtheler@tom:~/spack_psd$ spack info psd
AutotoolsPackage:   psd

Description:
     PSD (Parallel finite element Solver for continuum Dynamics) is a high-
    performance finite element solver for solid mechanics applications.
    Designed for largescale simulations, it is capable of solving problems
    with large numberof unknowns in both static & dynamic regimes, with
    support for linear and nonlinear models

Homepage: https://mohd-afeef-badri.github.io/psd

Preferred version:
    main    [git] https://github.com/mohd-afeef-badri/psd.git on branch main

Safe versions:
    main    [git] https://github.com/mohd-afeef-badri/psd.git on branch main

Deprecated versions:
    None

Variants:
    build_system [autotools]        autotools
        Build systems supported by the package

Dependencies:
    autoconf                            build

    automake                            build

    cxx                                 build

    freefempsd+mpi+petsc~superlu        build, link

    gmake                               build
      when  build_system=autotools

    gnuconfig                           build
      when  build_system=autotools target=ppc64le:

    gnuconfig                           build
      when  build_system=autotools target=aarch64:

    gnuconfig                           build
      when  build_system=autotools target=riscv64:

    libtool                             build

    m4                                  build

    mgispsd                             build, link

    petsc+hpddm+mpi+mumps               build, link

    tfel                                build, link

Licenses:
    Apache-2.0

gtheler@tom:~/spack_psd$
```

Then into the tutorial

```
gtheler@tom:~/run/psd$ PSD_PreProcess -problem linear_elasticity -dimension 2 -bodyforceconditions 1 -dirichletconditions 1 -postprocess u
===================================================================
                        ___               __
                      / __ \   _____   __/  /
                     / /_/ /  / ___/ / __  /
                    / ___ /  (__  ) / /_/ /
                   /_/      /____/  \___ /

      Parallel finite element Solver for continuum Dynamics

===================================================================

 building Main.edp ..................................... Done
 building MeshAndFeSpace.edp ............... Done
 building Macros.edp ............... Done
 building FemParameters.edp ............... Done
 building OtherParameters.edp ............... Done
 building ControlParameters.edp ............... Done
 building VariationalFormulations.edp .................. Done
 building LinearFormBuilderAndSolver.edp ............... Done
 building PostProcessor.edp ............................ Done

===================================================================

 PSD solver is now ready to run.

 For a simulation with N number of processes run your solver with

     PSD_Solve -np N Main.edp

 For a sequential simulation run your solver with

     PSD_Solve_Seq Main.edp

===================================================================
gtheler@tom:~/run/psd$ PSD_Solve -np 4 Main.edp -mesh  ./bar.msh -v 0
'/home/gtheler/spack/opt/spack/linux-skylake/openmpi-5.0.9-habcjpoz5sosphjxageq4eubw427pgks/bin/mpiexec' --oversubscribe -np 4 FreeFem++-mpi -nw 'Main.edp' -mesh ./bar.msh -v 0
  current line = 22 mpirank 3 / 4
  current line = 22 mpirank 2 / 4
  current line = 22 mpirank 1 / 4

Load error: parmetis
         fail:
 dlerror : parmetis.so: cannot open shared object file: No such file or directory
list prefix: './' '/home/gtheler/spack/opt/spack/linux-skylake/freefempsd-4.15-fbut46mtffbv3ayx4da7du4xjcbgyj5y/lib/ff++/4.15/lib/mpi/' '/home/gtheler/spack/opt/spack/linux-skylake/psd-main-uulo4hjj6ebwydr7jtil4ahblx4jvymr/include/psd/' '' list suffix: '' , '.so'
  current line = 22 mpirank 0 / 4
Load error : parmetis
        line number :22, parmetis
error Load error : parmetis
        line number :22, parmetis
 code = 2 mpirank: 0
WARNING! There are options you set that were not used!
WARNING! could be spelling mistake, etc!
There are 3 unused database options. They are:
Option left: name:-mesh value: ./bar.msh source: command line
Option left: name:-nw value: Main.edp source: command line
Option left: name:-v value: 0 source: command line
--------------------------------------------------------------------------
prterun detected that one or more processes exited with non-zero status,
thus causing the job to be terminated. The first process to do so was:
   Process name: [prterun-tom-985602@1,2]
   Exit code:    2
--------------------------------------------------------------------------
gtheler@tom:~/run/psd$
```
