### 2026-02-20 mohd-afeef-badri/psd#56 @mohd-afeef-badri
@gtheler I don't understand the issue well, in my understanding PETSc https://packages.spack.io/package.html?name=petsc on spack comes with Parmetis, which somehow does not seem to be the case. Can you try again please, I have explicitly mentioned Parmetis to be compiled in the spack repository for PSD, you will need to hit pull.

thank you again for all your help and patience
