### 2026-02-23 mohd-afeef-badri/psd#56 @gtheler
```
gtheler@tom:~$ cd spack_psd/
gtheler@tom:~/spack_psd$ git pull
remote: Enumerating objects: 13, done.
remote: Total 7 (delta 2), reused 0 (delta 0), pack-reused 0 (from 0)
Unpacking objects: 100% (7/7), 1.18 KiB | 606.00 KiB/s, done.
From https://github.com/mohd-afeef-badri/spack_psd
   85bbe07..f8f3d31  main       -> origin/main
Updating 85bbe07..f8f3d31
Fast-forward
 spack_repo/psd/packages/psd/package.py | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)
gtheler@tom:~/spack_psd$ cd ..
gtheler@tom:~$ cd spack
gtheler@tom:~/spack$ cd share/
gtheler@tom:~/spack/share$ cd spack/
gtheler@tom:~/spack/share/spack$ source setup-env.sh
cdgtheler@tom:~/spack/share/spack$ cd ..
gtheler@tom:~/spack/share$ cd ..
gtheler@tom:~/spack$ cd ..
gtheler@tom:~$ spack repo add spack_psd/spack_repo/psd/
==> Error: A repository with the name 'psd' already exists.
gtheler@tom:~$ spack install psd
==> Error: No such variant 'parmetis' in package petsc
gtheler@tom:~$
```
