### 2026-02-27 #9411 @nrichart
@diehlpk, @mohd-afeef-badri yes this answer my question, for me the paper is good and for the code I did manage to install it on my laptop and run tutorials. I just want to test an installation on a HPC cluster and run a bigger test case with the provided instruction above.
I did not manage to do it yet by lack of time but I should have time next week to do it.
