### 2026-03-19 #9411 @nrichart
@diehlpk I add issues compiling the code due to missing dependencies (so not due to the code itself)
I still have to run the tutorials
