### 2026-03-26 mohd-afeef-badri/psd#67 @nrichart
Running on a slurm cluster
Hello,

I am trying to run the tutorials on a slurm cluster where I should use `srun` to run mpi codes instead of mpirun
The issue is that `PSD_Solve` does the wrapping. Can I specify that I need to use srun or can I avoid the PSD_Solve wrapper

Thank you in advance

(This is related to  https://github.com/openjournals/joss-reviews/issues/9411)
Sorry for the delays in my tests

The software stack on the cluster uses spack that deploys a fake mpirun in openmpi with slurm support

```
$HOME/psd/PSDinstall/bin/PSD_Solve -np 4 Main.edp -v 0 -ns -nw
'/ssoft/spack/pinot-noir/jed/v2/spack/opt/spack/linux-rhel9-icelake/gcc-13.2.0/openmpi-5.0.3-i7bson2onqxzu5lpmymxfucij6wb5tbf/bin/mpirun' -np 4 /home/richart/psd/PSDinstall/bin/FreeFem++-mpi -nw 'Main.edp' -v 0 -ns
This version of Spack (openmpi ~legacylaunchers schedulers=slurm)
is installed without the mpiexec/mpirun commands to prevent
unintended performance issues. See https://github.com/spack/spack/pull/10340
for more details.
If you understand the potential consequences of a misconfigured mpirun, you can
use spack to install 'openmpi+legacylaunchers' to restore the executables.
Otherwise, use srun to launch your MPI executables.
```
