### 2026-04-06 #9411 @diehlpk
> [@diehlpk](https://github.com/diehlpk) I add issues compiling the code due to missing dependencies (so not due to the code itself) I still have to run the tutorials

@nrichart could you run the tutorials?
