### 2026-04-11 mohd-afeef-badri/psd#67 @mohd-afeef-badri
@nrichart sorry for the delayed response.

A quick  hacky way to run the code with srun would be to edit the file `$HOME/psd/PSDinstall/bin/PSD_Solve` you will need to add

```
mpirun="srun"
mpi_run_option=""
```

after the line
```
if [ -n "--oversubscribe"  ] ;then mpi_run_option="--oversubscribe"  ; fi
```

Basically this will override the `/ssoft/spack/pinot-noir/jed/v2/spack/opt/spack/linux-rhel9-icelake/gcc-13.2.0/openmpi-5.0.3-i7bson2onqxzu5lpmymxfucij6wb5tbf/bin/mpirun`  with `srun` and remove `--oversubscribe` as it is not needed with `srun` .

If this does not work.
Correct way to do it at the installation level would be to use `--with-MPIRUN=srun`  during the configure phase https://github.com/mohd-afeef-badri/psd/blob/a48ce2f31f86b520d2519cd9e1ff3be29f0dc9d9/configure.ac#L324

Please let me know if this resolves the issue. Further, I will add notes about this in the documentation.
