### 2026-04-16 mohd-afeef-badri/psd#67 @nrichart
@mohd-afeef-badri `--with-MPIRUN=` works thank you
I had to use `PSD_Solve -n <nrpoc>` instead of `-np`

In fact passing the srun option to PSD_Solve pass them to properly, but I did not find anything in the documentation.
