### 2026-04-17 mohd-afeef-badri/psd#67 @nrichart
I am trying to run the Linear elastic 2D case and it fails

```
$ PSD_Solve -n 4 -p infiniband --time 1:0:0 Main.edp -mesh ./../Meshes/2D/bar.msh -v 0
'/usr/bin/srun' -n 4 -p infiniband --time 1:0:0 /home/richart/psd/PSDinstall/bin/FreeFem++-mpi -nw 'Main.edp' -mesh ./../Meshes/2D/bar.msh -v 0
WARNING! There are options you set that were not used!
WARNING! could be spelling mistake, etc!
There are 3 unused database options. They are:
Option left: name:-mesh value: ./../Meshes/2D/bar.msh source: command line
Option left: name:-nw value: Main.edp source: command line
Option left: name:-v value: 0 source: command line
```
