### 2026-04-22 #9411 @nrichart
@diehlpk, my first issue running on a cluster was solved, but I reached another issue mentioned in #67. It might be a user error but I am a bit stuck
